+++ tests/test_app.py
@@ -13,3405 +13,31 @@ from werkzeug.security import generate_password_hash # Keep for AppTestCase stru
 # from datetime import datetime, timedelta ## Kept above
 # from werkzeug.security import generate_password_hash ## Kept above
 
-class AppTestCase(unittest.TestCase):
+# Import AppTestCase from test_base
+from tests.test_base import AppTestCase
 
-    @classmethod
-    def setUpClass(cls):
-        # Configuration that applies to the entire test class
-        # app.config['TESTING'] = True ## app IS NOT AVAILABLE
-        # app.config['WTF_CSRF_ENABLED'] = False ## app IS NOT AVAILABLE
-        # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:' ## app IS NOT AVAILABLE
-        # app.config['SECRET_KEY'] = 'test-secret-key' ## app IS NOT AVAILABLE
-        # To prevent real socketio connections during tests if not using test_client's context
-        # However, for `socketio.emit` testing, we often mock it anyway.
-        # app.config['SOCKETIO_MESSAGE_QUEUE'] = None ## app IS NOT AVAILABLE
-        # with app.app_context():
-            # db.create_all() # Create tables once per class
-        pass # Simplified for timeout debugging
+# TestFriendPostNotifications, TestDiscoverPageViews, and TestRecommendationAPI
+# have been moved to their respective files.
 
-    @classmethod
-    def tearDownClass(cls):
-        # with app.app_context():
-            # db.drop_all() # Drop tables once after all tests in the class
-        pass # Simplified for timeout debugging
+# TestPersonalizedFeedAPI, TestOnThisDayPage, TestOnThisDayAPI, and TestTrendingHashtags
+# have been moved to their respective files.
 
-    def setUp(self):
-        """Set up for each test."""
-        # self.client = app.test_client() ## app IS NOT AVAILABLE
-        # with app.app_context():
-            # Ensure tables are clean before setting up users for this test
-            # self._clean_tables_for_setup()
-            # self._setup_base_users()
-        pass # Simplified for timeout debugging
+# TestUserFeedAPI, TestUserStatsAPI, and TestUserStatus
+# have been moved to their respective files.
 
-    def _clean_tables_for_setup(self):
-        """Clears data from tables before each test's user setup."""
-        # This is an attempt to ensure a truly clean state if tearDown isn't enough
-        # or if running tests out of order / individually in some environments.
-        # db.session.remove() # Ensure session is clean before delete operations ## db IS NOT AVAILABLE
-        # for table in reversed(db.metadata.sorted_tables): ## db IS NOT AVAILABLE
-        #     db.session.execute(table.delete()) ## db IS NOT AVAILABLE
-        # db.session.commit() ## db IS NOT AVAILABLE
-        pass # Simplified for timeout debugging
+# TestLiveActivityFeed, TestFileSharing, TestSeriesFeature and _create_db_user_activity helper
+# have been moved to their respective files.
 
+# TestCommentAPI, TestPollAPI, TestCollaborativeEditing
+# have been moved to their respective files.
 
-    def tearDown(self):
-        """Executed after each test."""
-        # with app.app_context():
-            # db.session.remove() # Remove session first ## db IS NOT AVAILABLE
-            # Selectively delete data to speed up vs db.drop_all()
-            # for table in reversed(db.metadata.sorted_tables): ## db IS NOT AVAILABLE
-            #     db.session.execute(table.delete()) ## db IS NOT AVAILABLE
-            # db.session.commit() ## db IS NOT AVAILABLE
-        pass # Simplified for timeout debugging
-
-        # Clean up shared files directory after each test if it exists
-        # shared_files_folder = app.config.get('SHARED_FILES_UPLOAD_FOLDER') ## app IS NOT AVAILABLE
-        shared_files_folder = None # temp replacement
-        if shared_files_folder and os.path.exists(shared_files_folder):
-            for filename in os.listdir(shared_files_folder):
-                file_path = os.path.join(shared_files_folder, filename)
-                try:
-                    if os.path.isfile(file_path) or os.path.islink(file_path):
-                        os.unlink(file_path)
-                except Exception as e:
-                    print(f"Failed to delete {file_path}. Reason: {e}")
-
-
-    def _setup_base_users(self):
-        # _clean_tables_for_setup (called in setUp) should have handled table clearing.
-        # User.query.delete()
-        # db.session.commit()
-        # Using instance variables for user objects to make them accessible in tests
-        self.user1 = User(username='testuser1', email='test1@example.com', password_hash=generate_password_hash('password'))
-        self.user2 = User(username='testuser2', email='test2@example.com', password_hash=generate_password_hash('password'))
-        self.user3 = User(username='testuser3', email='test3@example.com', password_hash=generate_password_hash('password'))
-        db.session.add_all([self.user1, self.user2, self.user3])
-        db.session.commit()
-        # Store IDs for later use, helpful if objects become detached or for clarity
-        self.user1_id = self.user1.id
-        self.user2_id = self.user2.id
-        self.user3_id = self.user3.id
-
-    def _create_friendship(self, user1_id, user2_id, status='accepted'):
-        friendship = Friendship(user_id=user1_id, friend_id=user2_id, status=status)
-        db.session.add(friendship)
-        db.session.commit()
-        return friendship
-
-    def _create_db_post(self, user_id, title="Test Post", content="Test Content", timestamp=None):
-        post = Post(user_id=user_id, title=title, content=content, timestamp=timestamp or datetime.utcnow())
-        db.session.add(post)
-        db.session.commit()
-        return post
-
-    def _make_post_via_route(self, username, password, title="Test Post", content="Test Content", hashtags=""):
-        self.login(username, password)
-        response = self.client.post('/blog/create', data=dict(
-            title=title,
-            content=content,
-            hashtags=hashtags
-        ), follow_redirects=True)
-        self.logout() # Logout after post creation to keep session state clean for next login
-        return response
-
-    def login(self, username, password):
-        return self.client.post('/login', data=dict(
-            username=username,
-            password=password
-        ), follow_redirects=True)
-
-    def logout(self):
-        return self.client.get('/logout', follow_redirects=True)
-
-    def _create_db_message(self, sender_id, receiver_id, content, timestamp=None, is_read=False):
-        # This helper now operates within an app_context implicitly if called from a test method that has it.
-        # If called from setUpClass or outside a request context, ensure app_context.
-        msg = Message(
-            sender_id=sender_id,
-            receiver_id=receiver_id,
-            content=content,
-            timestamp=timestamp or datetime.utcnow(),
-            is_read=is_read
-        )
-        db.session.add(msg)
-        db.session.commit()
-        return msg
-
-    # Example existing tests (shortened for brevity)
-    @patch('app.socketio.emit')
-    def test_send_message_real_time(self, mock_socketio_emit):
-        # This test seems to be about direct messages, will keep it as is.
-        # My new tests will focus on FriendPostNotification.
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            # ... (rest of the existing test) ...
-            self.logout()
-
-    def test_view_conversation_marks_read(self):
-        with app.app_context():
-            msg = self._create_db_message(sender_id=self.user1_id, receiver_id=self.user2_id, content="Test", is_read=False)
-            # ... (rest of the existing test) ...
-            self.logout()
-
-    def test_inbox_route_data(self):
-        with app.app_context():
-            # ... (rest of the existing test) ...
-            self.logout()
-
-
-class TestFriendPostNotifications(AppTestCase): # Inherit from AppTestCase for setup
-    pass # All tests commented out for now to isolate timeout issues
-
-    # @patch('app.socketio.emit') # Patch socketio.emit from the app instance
-    # def test_notification_creation_and_socketio_emit(self, mock_socketio_emit):
-    #     with app.app_context():
-    #         # 1. User A and User B are friends.
-    #         self._create_friendship(self.user1_id, self.user2_id, status='accepted')
-    #
-    #         # 2. User A creates a new post.
-    #         post_title = "User A's Exciting Post"
-    #         self._make_post_via_route(self.user1.username, 'password', title=post_title, content="Content here")
-    #
-    #         # Retrieve the post created by User A
-    #         created_post = Post.query.filter_by(user_id=self.user1_id, title=post_title).first()
-    #         self.assertIsNotNone(created_post)
-    #
-    #         # 3. Assert that a FriendPostNotification record is created for User B
-    #         notification_for_b = FriendPostNotification.query.filter_by(
-    #             user_id=self.user2_id,
-    #             post_id=created_post.id,
-    #             poster_id=self.user1_id
-    #         ).first()
-    #         self.assertIsNotNone(notification_for_b)
-    #         self.assertFalse(notification_for_b.is_read)
-    #
-    #         # 4. Assert that no notification is created for User A
-    #         notification_for_a = FriendPostNotification.query.filter_by(
-    #             user_id=self.user1_id,
-    #             post_id=created_post.id
-    #         ).first()
-    #         self.assertIsNone(notification_for_a)
-    #
-    #         # 5. Assert that if User C is not friends with User A, User C does not receive a notification.
-    #         notification_for_c = FriendPostNotification.query.filter_by(
-    #             user_id=self.user3_id,
-    #             post_id=created_post.id
-    #         ).first()
-    #         self.assertIsNone(notification_for_c)
-    #
-    #         # 6. Assert socketio.emit was called for User B
-    #         expected_socket_payload = {
-    #             'notification_id': notification_for_b.id,
-    #             'post_id': created_post.id,
-    #             'post_title': created_post.title,
-    #             'poster_username': self.user1.username,
-    #             'timestamp': ANY # Timestamps can be tricky, mock with ANY or compare with tolerance
-    #         }
-    #         # We need to compare timestamp more carefully if not using ANY
-    #         # For now, ANY is simpler. If using specific timestamp, ensure it matches notification_for_b.timestamp.isoformat()
-    #
-    #         mock_socketio_emit.assert_any_call(
-    #             'new_friend_post',
-    #             expected_socket_payload,
-    #             room=f'user_{self.user2_id}'
-    #         )
-    #         # Check it wasn't called for user A or C for this specific post
-    #         # This is harder to assert directly without more complex call tracking if emit is called for other reasons
-    #         # The DB checks largely cover this.
-    #
-    # def test_view_friend_post_notifications_page(self):
-    #     with app.app_context():
-    #         # User1 and User2 are friends. User1 posts. User2 gets a notification.
-    #         self._create_friendship(self.user1_id, self.user2_id)
-    #         post1_by_user1 = self._create_db_post(user_id=self.user1_id, title="Post 1 by User1", timestamp=datetime.utcnow() - timedelta(minutes=10))
-    #         # Manually create notification as if post route was hit by user1
-    #         notif1_for_user2 = FriendPostNotification(user_id=self.user2_id, post_id=post1_by_user1.id, poster_id=self.user1_id, timestamp=post1_by_user1.timestamp)
-    #
-    #         # User3 and User2 are friends. User3 posts. User2 gets another notification (newer).
-    #         self._create_friendship(self.user3_id, self.user2_id)
-    #         post2_by_user3 = self._create_db_post(user_id=self.user3_id, title="Post 2 by User3", timestamp=datetime.utcnow() - timedelta(minutes=5))
-    #         notif2_for_user2 = FriendPostNotification(user_id=self.user2_id, post_id=post2_by_user3.id, poster_id=self.user3_id, timestamp=post2_by_user3.timestamp)
-    #
-    #         db.session.add_all([notif1_for_user2, notif2_for_user2])
-    #         db.session.commit()
-    #
-    #         self.login(self.user2.username, 'password')
-    #         response = self.client.get('/friend_post_notifications')
-    #         self.assertEqual(response.status_code, 200)
-    #         response_data = response.get_data(as_text=True)
-    #
-    #         self.assertIn(self.user3.username, response_data) # Poster of newer notification
-    #         self.assertIn(post2_by_user3.title, response_data)
-    #         self.assertIn(self.user1.username, response_data) # Poster of older notification
-    #         self.assertIn(post1_by_user1.title, response_data)
-    #
-    #         # Assert order (newer notification from user3 appears before older from user1)
-    #         self.assertTrue(response_data.find(post2_by_user3.title) < response_data.find(post1_by_user1.title))
-    #         self.logout()
-    #
-    # def test_mark_one_notification_as_read(self):
-    #     with app.app_context():
-    #         self._create_friendship(self.user1_id, self.user2_id)
-    #         post_by_user1 = self._create_db_post(user_id=self.user1_id)
-    #         notification = FriendPostNotification(user_id=self.user2_id, post_id=post_by_user1.id, poster_id=self.user1_id, is_read=False)
-    #         db.session.add(notification)
-    #         db.session.commit()
-    #         notification_id = notification.id
-    #
-    #         self.assertFalse(FriendPostNotification.query.get(notification_id).is_read)
-    #
-    #         # User2 (owner) marks as read
-    #         self.login(self.user2.username, 'password')
-    #         response = self.client.post(f'/friend_post_notifications/mark_as_read/{notification_id}')
-    #         self.assertEqual(response.status_code, 200)
-    #         self.assertEqual(response.json, {'status': 'success', 'message': 'Notification marked as read.'})
-    #         self.assertTrue(FriendPostNotification.query.get(notification_id).is_read)
-    #         self.logout()
-    #
-    #         # User3 (not owner) tries to mark as read
-    #         # First, set it back to unread for this part of the test
-    #         notification_db = FriendPostNotification.query.get(notification_id)
-    #         notification_db.is_read = False
-    #         db.session.commit()
-    #         self.assertFalse(FriendPostNotification.query.get(notification_id).is_read)
-    #
-    #         self.login(self.user3.username, 'password')
-    #         response = self.client.post(f'/friend_post_notifications/mark_as_read/{notification_id}')
-    #         self.assertEqual(response.status_code, 403) # Forbidden
-    #         self.assertEqual(response.json, {'status': 'error', 'message': 'Unauthorized.'})
-    #         self.assertFalse(FriendPostNotification.query.get(notification_id).is_read) # Still false
-    #         self.logout()
-    #
-    #         # Test non-existent notification
-    #         self.login(self.user2.username, 'password')
-    #         response = self.client.post(f'/friend_post_notifications/mark_as_read/99999')
-    #         self.assertEqual(response.status_code, 404)
-    #         self.assertEqual(response.json, {'status': 'error', 'message': 'Notification not found.'})
-    #         self.logout()
-    #
-    #
-    # def test_mark_all_notifications_as_read(self):
-    #     with app.app_context():
-    #         self._create_friendship(self.user1_id, self.user2_id)
-    #         post1 = self._create_db_post(user_id=self.user1_id, title="Post1")
-    #         post2 = self._create_db_post(user_id=self.user1_id, title="Post2")
-    #
-    #         notif1 = FriendPostNotification(user_id=self.user2_id, post_id=post1.id, poster_id=self.user1_id, is_read=False)
-    #         notif2 = FriendPostNotification(user_id=self.user2_id, post_id=post2.id, poster_id=self.user1_id, is_read=False)
-    #         # Notification for another user (user3) - should not be affected
-    #         notif_for_user3 = FriendPostNotification(user_id=self.user3_id, post_id=post1.id, poster_id=self.user1_id, is_read=False)
-    #
-    #         db.session.add_all([notif1, notif2, notif_for_user3])
-    #         db.session.commit()
-    #         notif1_id, notif2_id, notif3_id = notif1.id, notif2.id, notif_for_user3.id
-    #
-    #
-    #         self.login(self.user2.username, 'password')
-    #         response = self.client.post('/friend_post_notifications/mark_all_as_read')
-    #         self.assertEqual(response.status_code, 200)
-    #         self.assertEqual(response.json, {'status': 'success', 'message': 'All friend post notifications marked as read.'})
-    #
-    #         self.assertTrue(FriendPostNotification.query.get(notif1_id).is_read)
-    #         self.assertTrue(FriendPostNotification.query.get(notif2_id).is_read)
-    #         self.assertFalse(FriendPostNotification.query.get(notif3_id).is_read) # User3's notification untouched
-    #         self.logout()
-    #
-    #         # Test when no unread notifications exist for the user
-    #         self.login(self.user2.username, 'password') # user2's notifs are now read
-    #         response = self.client.post('/friend_post_notifications/mark_all_as_read')
-    #         self.assertEqual(response.status_code, 200)
-    #         self.assertEqual(response.json, {'status': 'success', 'message': 'No unread friend post notifications.'})
-    #         self.logout()
-
-
-class TestDiscoverPageViews(AppTestCase):
-
-    @patch('app.get_personalized_feed_posts')
-    def test_discover_page_shows_recommendation_reasons(self, mock_get_personalized_feed_posts):
-        with app.app_context():
-            # Setup: Login as a user
-            self.login(self.user1.username, 'password')
-
-            # Mocking get_personalized_feed_posts
-            # Create a mock author object that behaves like a User model instance for the template
-            mock_author = unittest.mock.MagicMock(spec=User)
-            mock_author.username = 'author_username'
-
-            # Create a mock post object that behaves like a Post model instance for the template
-            mock_post = unittest.mock.MagicMock(spec=Post)
-            mock_post.id = 123
-            mock_post.title = "Mocked Post Title"
-            # Ensure content is not None for slicing in template (post.content[:200])
-            mock_post.content = "Mocked post content here that is long enough."
-            mock_post.author = mock_author
-            # Add other attributes that might be accessed if post.to_dict() was called, or by template directly
-            mock_post.user_id = self.user2_id # Assuming user2 might be an author
-            mock_post.timestamp = datetime.utcnow()
-            mock_post.comments = [] # For len(post.comments) if used
-            mock_post.likes = []    # For len(post.likes) if used
-            mock_post.reviews = []  # For len(post.reviews) if used
-            mock_post.hashtags = ""
-            mock_post.is_featured = False
-            mock_post.featured_at = None
-            mock_post.last_edited = None
-
-
-            mock_reason = "Test reason for this post."
-            # The function is expected to return a list of (Post, reason_string) tuples
-            mock_get_personalized_feed_posts.return_value = [(mock_post, mock_reason)]
-
-            # Execution: Make a GET request to /discover
-            response = self.client.get('/discover')
-
-            # Assertions
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            # Check for the reason string
-            self.assertIn(f"Recommended because: {mock_reason}", response_data)
-            # Check for post details
-            self.assertIn(mock_post.title, response_data)
-            self.assertIn(mock_post.author.username, response_data)
-            # Check a snippet of content if it's displayed
-            self.assertIn(mock_post.content[:50], response_data)
-
-            # Assert that the mock was called correctly
-            mock_get_personalized_feed_posts.assert_called_once_with(self.user1_id, limit=15)
-
-            self.logout()
-
-class TestRecommendationAPI(AppTestCase):
-
-    def _create_db_group(self, creator_id, name="Test Group", description="A group for testing"):
-        group = Group(name=name, description=description, creator_id=creator_id)
-        db.session.add(group)
-        db.session.commit()
-        return group
-
-    def _create_db_event(self, user_id, title="Test Event", description="An event for testing", date="2024-12-31", time="18:00", location="Test Location"):
-        event = Event(title=title, description=description, date=date, time=time, location=location, user_id=user_id)
-        db.session.add(event)
-        db.session.commit()
-        return event
-
-    def _create_db_poll(self, user_id, question="Test Poll?", options_texts=None):
-        if options_texts is None:
-            options_texts = ["Option 1", "Option 2"]
-        poll = Poll(question=question, user_id=user_id)
-        db.session.add(poll)
-        db.session.flush() # So poll gets an ID before adding options
-        for text in options_texts:
-            option = PollOption(text=text, poll_id=poll.id)
-            db.session.add(option)
-        db.session.commit()
-        return poll
-
-    def setUp(self):
-        super().setUp() # Call parent setUp to get base users (self.user1, self.user2, self.user3)
-        # self.user1 is the target user for recommendations
-        # self.user2 will create content
-        # self.user3 is the "lonely" user
-
-        # Create some content by user2 that user1 might be recommended
-        self.post_by_user2 = self._create_db_post(user_id=self.user2_id, title="User2's Post")
-        self.group_by_user2 = self._create_db_group(creator_id=self.user2_id, name="User2's Group")
-        self.event_by_user2 = self._create_db_event(user_id=self.user2_id, title="User2's Event")
-        self.poll_by_user2 = self._create_db_poll(user_id=self.user2_id, question="User2's Poll?")
-
-        # User1 joins a different group (not by user2) to test suggest_groups_to_join logic (won't recommend this one)
-        # self.other_group_user1_member_of = self._create_db_group(creator_id=self.user3_id, name="Other Group")
-        # self.other_group_user1_member_of.members.append(self.user1)
-        # db.session.commit()
-
-
-    def test_get_recommendations_success(self):
-        with app.app_context():
-            response = self.client.get(f'/api/recommendations?user_id={self.user1_id}')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.content_type, 'application/json')
-
-            data = json.loads(response.data)
-
-            self.assertIn('user_id', data)
-            self.assertEqual(data['user_id'], self.user1_id)
-
-            self.assertIn('suggested_posts', data)
-            self.assertIsInstance(data['suggested_posts'], list)
-            if data['suggested_posts']:
-                post = data['suggested_posts'][0]
-                self.assertIn('id', post)
-                self.assertIn('title', post)
-                self.assertIn('author_username', post)
-
-            self.assertIn('suggested_groups', data)
-            self.assertIsInstance(data['suggested_groups'], list)
-            if data['suggested_groups']:
-                group = data['suggested_groups'][0]
-                self.assertIn('id', group)
-                self.assertIn('name', group)
-                self.assertIn('creator_username', group)
-
-            self.assertIn('suggested_events', data)
-            self.assertIsInstance(data['suggested_events'], list)
-            if data['suggested_events']:
-                event = data['suggested_events'][0]
-                self.assertIn('id', event)
-                self.assertIn('title', event)
-                self.assertIn('organizer_username', event)
-
-            self.assertIn('suggested_users_to_follow', data)
-            self.assertIsInstance(data['suggested_users_to_follow'], list)
-            if data['suggested_users_to_follow']:
-                user = data['suggested_users_to_follow'][0]
-                self.assertIn('id', user)
-                self.assertIn('username', user)
-
-            self.assertIn('suggested_polls_to_vote', data)
-            self.assertIsInstance(data['suggested_polls_to_vote'], list)
-            if data['suggested_polls_to_vote']:
-                poll = data['suggested_polls_to_vote'][0]
-                self.assertIn('id', poll)
-                self.assertIn('question', poll)
-                self.assertIn('author_username', poll)
-                self.assertIn('options', poll)
-                self.assertIsInstance(poll['options'], list)
-                if poll['options']:
-                    option = poll['options'][0]
-                    self.assertIn('id', option)
-                    self.assertIn('text', option)
-                    self.assertIn('vote_count', option)
-
-    def test_get_recommendations_invalid_user_id(self):
-        with app.app_context():
-            response = self.client.get('/api/recommendations?user_id=99999')
-            self.assertEqual(response.status_code, 404)
-            data = json.loads(response.data)
-            self.assertIn('message', data)
-            # The message might be "User 99999 not found" or "User not found"
-            # Let's check if "not found" is in the message for robustness
-            self.assertTrue('not found' in data['message'].lower())
-
-
-    def test_get_recommendations_missing_user_id(self):
-        with app.app_context():
-            response = self.client.get('/api/recommendations')
-            self.assertEqual(response.status_code, 400)
-            data = json.loads(response.data)
-            self.assertIn('message', data)
-            # Example: {'message': {'user_id': 'User ID is required and must be an integer.'}}
-            # Exact message depends on reqparse error formatting
-            self.assertIn('user_id', data['message'])
-            self.assertTrue('required' in data['message']['user_id'].lower())
-
-
-    def test_get_recommendations_no_suggestions(self):
-        # self.user3 is set up by AppTestCase.setUp -> _setup_base_users()
-        # It has no specific content or interactions created in this class's setUp,
-        # so it should get minimal to no recommendations.
-        with app.app_context():
-            response = self.client.get(f'/api/recommendations?user_id={self.user3_id}')
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-
-            self.assertEqual(data['user_id'], self.user3_id)
-            self.assertEqual(data['suggested_posts'], [])
-            self.assertEqual(data['suggested_groups'], [])
-            self.assertEqual(data['suggested_events'], [])
-            # user3 might be recommended user1 and user2 if the suggestion logic is simple
-            # For now, let's assert it's a list. More specific checks depend on recommendation logic.
-            self.assertIsInstance(data['suggested_users_to_follow'], list)
-            self.assertEqual(data['suggested_polls_to_vote'], [])
-
-    # _get_jwt_token moved to AppTestCase
-
-    # Helpers for creating likes, comments, RSVPs, votes
-    def _create_db_like(self, user_id, post_id, timestamp=None):
-        from models import Like # Import here to avoid circular dependency if models import app
-        like = Like(user_id=user_id, post_id=post_id, timestamp=timestamp or datetime.utcnow())
-        db.session.add(like)
-        db.session.commit()
-        return like
-
-    def _create_db_comment(self, user_id, post_id, content="Test comment", timestamp=None):
-        from models import Comment # Import here
-        comment = Comment(user_id=user_id, post_id=post_id, content=content, timestamp=timestamp or datetime.utcnow())
-        db.session.add(comment)
-        db.session.commit()
-        return comment
-
-    def _create_db_event_rsvp(self, user_id, event_id, status="Attending", timestamp=None):
-        from models import EventRSVP # Import here
-        rsvp = EventRSVP(user_id=user_id, event_id=event_id, status=status, timestamp=timestamp or datetime.utcnow())
-        db.session.add(rsvp)
-        db.session.commit()
-        return rsvp
-
-    def _create_db_poll_vote(self, user_id, poll_id, poll_option_id, timestamp=None):
-        from models import PollVote # Import here
-        vote = PollVote(user_id=user_id, poll_id=poll_id, poll_option_id=poll_option_id, timestamp=timestamp or datetime.utcnow())
-        db.session.add(vote)
-        db.session.commit()
-        return vote
-
-
-class TestPersonalizedFeedAPI(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # Users are created by AppTestCase's _setup_base_users: self.user1, self.user2, self.user3
-        # self.user1_id, self.user2_id, self.user3_id
-
-    def test_personalized_feed_unauthorized(self):
-        with app.app_context():
-            response = self.client.get('/api/personalized-feed')
-            self.assertEqual(response.status_code, 401) # JWT errors are usually 401 or 422 if malformed
-            # Flask-JWT-Extended typically returns 401 for missing token
-            data = json.loads(response.data)
-            self.assertIn('msg', data) # Default message key for flask-jwt-extended
-            self.assertEqual(data['msg'], 'Missing Authorization Header')
-
-
-class TestOnThisDayPage(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.test_user = self.user1
-        self.fixed_today = datetime(2023, 10, 26, 12, 0, 0) # Fixed date for testing - Oct 26
-
-        # Create a helper for events, similar to _create_db_post
-        self.post_target_correct = self._create_db_post(
-            user_id=self.test_user.id, title="Correct Web Post", content="Web content from Oct 26, 2022",
-            timestamp=datetime(2022, 10, 26, 10, 0, 0)
-        )
-        self.event_target_correct = self._create_db_event( # Use the helper defined below
-            user_id=self.test_user.id, title="Correct Web Event", date_str='2022-10-26',
-            description="Web event on Oct 26, 2022"
-        )
-        # For filtering tests
-        self.post_current_year_web = self._create_db_post(
-            user_id=self.test_user.id, title="Current Year Web Post",
-            timestamp=datetime(2023, 10, 26, 11, 0, 0)
-        )
-        self.event_different_day_web = self._create_db_event( # Use the helper
-            user_id=self.test_user.id, title="Different Day Web Event", date_str='2022-10-27'
-        )
-        db.session.commit() # Commit all created items
-
-    def _create_db_event(self, user_id, title, date_str, description="Test Event", time="12:00", location="Test Location", created_at=None):
-        event = Event(
-            user_id=user_id, title=title, description=description,
-            date=date_str, time=time, location=location,
-            created_at=created_at or datetime.utcnow()
-        )
-        db.session.add(event)
-        # db.session.commit() # Commit handled in setUp or test method after all creations
-        return event
-
-    def test_on_this_day_page_unauthorized(self):
-        with app.app_context():
-            response = self.client.get('/onthisday', follow_redirects=False)
-            self.assertEqual(response.status_code, 302)
-            self.assertTrue(response.location.endswith('/login')) # Check if ends with /login
-
-    @patch('app.datetime') # Patch datetime used in app.py for the route's internal logic
-    @patch('recommendations.datetime') # Patch datetime used in the recommendations.py function
-    def test_on_this_day_page_no_content(self, mock_reco_datetime, mock_app_datetime):
-        with app.app_context():
-            no_content_date = datetime(2023, 1, 1, 12, 0, 0)
-            mock_app_datetime.utcnow.return_value = no_content_date
-            mock_reco_datetime.utcnow.return_value = no_content_date
-            mock_reco_datetime.strptime = datetime.strptime # Ensure strptime is not mocked
-
-            self.login(self.test_user.username, 'password')
-            response = self.client.get('/onthisday')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            self.assertIn("No posts from this day in previous years.", response_data)
-            self.assertIn("No events from this day in previous years.", response_data)
-            # Check for the overall message if both are empty
-            self.assertIn("Nothing to show for 'On This Day' from previous years.", response_data)
-            self.logout()
-
-    @patch('app.datetime') # Patch datetime used in app.py for the route's internal logic
-    @patch('recommendations.datetime') # Patch datetime used in the recommendations.py function
-    def test_on_this_day_page_with_content_and_filtering(self, mock_reco_datetime, mock_app_datetime):
-        with app.app_context():
-            from flask import url_for # Import url_for within app_context for tests
-            mock_app_datetime.utcnow.return_value = self.fixed_today
-            mock_reco_datetime.utcnow.return_value = self.fixed_today
-            mock_reco_datetime.strptime = datetime.strptime
-
-            self.login(self.test_user.username, 'password')
-            response = self.client.get('/onthisday')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            # Check for correct content
-            self.assertIn(self.post_target_correct.title, response_data)
-            self.assertIn(self.event_target_correct.title, response_data)
-            self.assertIn(url_for('view_post', post_id=self.post_target_correct.id), response_data)
-            self.assertIn(url_for('view_event', event_id=self.event_target_correct.id), response_data)
-
-            # Check that filtered content is NOT present
-            self.assertNotIn(self.post_current_year_web.title, response_data)
-            self.assertNotIn(self.event_different_day_web.title, response_data)
-            self.logout()
-
-
-    def test_personalized_feed_success_and_structure(self):
-        with app.app_context():
-            # 1. Setup Data
-            # Friendships: user1 is friends with user2
-            self._create_friendship(self.user1_id, self.user2_id, status='accepted')
-
-            # Posts:
-            # Post by user3 (not friend), liked by user2 (friend of user1) -> should be recommended
-            post1_by_user3 = self._create_db_post(user_id=self.user3_id, title="Post by User3, Liked by User2", timestamp=datetime.utcnow() - timedelta(hours=1))
-            self._create_db_like(user_id=self.user2_id, post_id=post1_by_user3.id, timestamp=datetime.utcnow() - timedelta(minutes=30))
-
-            # Post by user2 (friend of user1) -> should be recommended (because friend created it, even if no other interaction)
-            # The current suggest_posts_to_read logic might not pick this up unless it's also interacted with by another friend or meets other criteria.
-            # For testing, let's ensure it's picked up. We might need to make user1 also friend with user3, and user3 likes user2's post.
-            # Let's simplify: user3 (not friend) posts, user2 (friend) comments.
-            post2_by_user3 = self._create_db_post(user_id=self.user3_id, title="Another Post by User3, Commented by User2", timestamp=datetime.utcnow() - timedelta(hours=3))
-            self._create_db_comment(user_id=self.user2_id, post_id=post2_by_user3.id, content="Friend comment", timestamp=datetime.utcnow() - timedelta(hours=2))
-
-            # Events:
-            # Event by user3, user2 (friend of user1) RSVP'd 'Attending' -> should be recommended
-            event1_by_user3 = self._create_db_event(user_id=self.user3_id, title="Event by User3, User2 Attending", date="2025-01-01")
-            event1_by_user3.created_at = datetime.utcnow() - timedelta(days=1) # Set created_at for sorting
-            db.session.commit()
-            self._create_db_event_rsvp(user_id=self.user2_id, event_id=event1_by_user3.id, status="Attending")
-
-            # Polls:
-            # Poll by user2 (friend of user1) -> should be recommended
-            poll1_by_user2 = self._create_db_poll(user_id=self.user2_id, question="Poll by User2 (Friend)?")
-            poll1_by_user2.created_at = datetime.utcnow() - timedelta(days=2) # Set created_at
-            db.session.commit()
-            # Add a vote from user3 to make it seem active, though suggest_polls_to_vote might pick it up just by being friend-created
-            option_for_poll1 = poll1_by_user2.options[0]
-            self._create_db_poll_vote(user_id=self.user3_id, poll_id=poll1_by_user2.id, poll_option_id=option_for_poll1.id)
-
-
-            # 2. Login as user1 and get token
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            # 3. Make request
-            response = self.client.get('/api/personalized-feed', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-            self.assertIn('feed_items', data)
-            feed_items = data['feed_items']
-            self.assertIsInstance(feed_items, list)
-
-            # We expect at least one of each type based on setup
-            self.assertTrue(len(feed_items) >= 3, f"Expected at least 3 items, got {len(feed_items)}")
-
-            found_post = False
-            found_event = False
-            found_poll = False
-
-            timestamps = []
-
-            for item in feed_items:
-                self.assertIn('type', item)
-                self.assertIn('id', item)
-                self.assertIn('timestamp', item)
-                self.assertIsNotNone(item['timestamp'])
-                timestamps.append(datetime.fromisoformat(item['timestamp'].replace('Z', ''))) # Handle Z if present
-
-                if item['type'] == 'post':
-                    found_post = True
-                    self.assertIn('title', item)
-                    self.assertIn('content', item)
-                    self.assertIn('author_username', item)
-                    self.assertIn('reason', item)
-                    if item['id'] == post1_by_user3.id:
-                        self.assertEqual(item['title'], "Post by User3, Liked by User2")
-                elif item['type'] == 'event':
-                    found_event = True
-                    self.assertIn('title', item)
-                    self.assertIn('description', item)
-                    self.assertIn('organizer_username', item)
-                    if item['id'] == event1_by_user3.id:
-                        self.assertEqual(item['title'], "Event by User3, User2 Attending")
-                elif item['type'] == 'poll':
-                    found_poll = True
-                    self.assertIn('question', item)
-                    self.assertIn('creator_username', item)
-                    self.assertIn('options', item)
-                    self.assertIsInstance(item['options'], list)
-                    if item['options']:
-                        self.assertIn('text', item['options'][0])
-                        self.assertIn('vote_count', item['options'][0])
-                    if item['id'] == poll1_by_user2.id:
-                         self.assertEqual(item['question'], "Poll by User2 (Friend)?")
-
-            self.assertTrue(found_post, "No post found in feed")
-            self.assertTrue(found_event, "No event found in feed")
-            self.assertTrue(found_poll, "No poll found in feed")
-
-            # Assert sorting by timestamp (descending)
-            for i in range(len(timestamps) - 1):
-                self.assertGreaterEqual(timestamps[i], timestamps[i+1], "Feed items are not sorted correctly by timestamp")
-
-    def test_personalized_feed_empty(self):
-        with app.app_context():
-            # user3 is set up but has no friends or interactions that would generate a feed for them based on current logic
-            token = self._get_jwt_token(self.user3.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.get('/api/personalized-feed', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-            self.assertIn('feed_items', data)
-            self.assertEqual(data['feed_items'], [])
-
-
-class TestTrendingHashtags(AppTestCase):
-
-    def test_update_trending_hashtags_logic(self):
-        with self.app.app_context():
-            # Clean up existing data
-            Post.query.delete()
-            TrendingHashtag.query.delete()
-            db.session.commit()
-
-            # Create sample posts
-            # self.user1_id is available from AppTestCase's _setup_base_users
-            db.session.add(Post(title="Post 1", content="Test", user_id=self.user1_id, hashtags="flask,python,web", timestamp=datetime.utcnow() - timedelta(days=1)))
-            db.session.add(Post(title="Post 2", content="Test", user_id=self.user1_id, hashtags="python,api,flask", timestamp=datetime.utcnow() - timedelta(days=2)))
-            db.session.add(Post(title="Post 3", content="Test", user_id=self.user1_id, hashtags="python,web", timestamp=datetime.utcnow() - timedelta(days=3))) # python:3, flask:2, web:2, api:1
-            db.session.add(Post(title="Post 4", content="Test", user_id=self.user1_id, hashtags="java,spring", timestamp=datetime.utcnow() - timedelta(days=1))) # java:1, spring:1
-            db.session.add(Post(title="Post 5", content="Test", user_id=self.user1_id, hashtags="python,old,web", timestamp=datetime.utcnow() - timedelta(days=10))) # This is old
-            db.session.commit()
-
-            # Call the function to update trending hashtags (defaults: top_n=10, since_days=7)
-            update_trending_hashtags()
-
-            updated_trends = TrendingHashtag.query.order_by(TrendingHashtag.rank.asc()).all()
-
-            self.assertTrue(len(updated_trends) > 0, "No trending hashtags were generated.")
-            self.assertTrue(len(updated_trends) <= 10, "More than 10 trending hashtags were generated.")
-
-            # Based on recent posts (Post 1, 2, 3, 4):
-            # python: 3 (from post 1, 2, 3)
-            # flask: 2 (from post 1, 2)
-            # web: 2 (from post 1, 3)
-            # api: 1 (from post 2)
-            # java: 1 (from post 4)
-            # spring: 1 (from post 4)
-            # Post 5's "python,old,web" should be ignored due to since_days=7 default
-
-            self.assertEqual(updated_trends[0].hashtag, "python")
-            self.assertEqual(updated_trends[0].rank, 1)
-            self.assertEqual(updated_trends[0].score, 3.0) # Should be 3 from recent posts
-
-            # The order of rank 2 can vary between flask and web as they both have score 2
-            # So, we check if the next two are flask and web in any order
-            rank2_hashtags = {updated_trends[1].hashtag, updated_trends[2].hashtag}
-            self.assertIn("flask", rank2_hashtags)
-            self.assertIn("web", rank2_hashtags)
-            self.assertEqual(updated_trends[1].rank, 2)
-            self.assertEqual(updated_trends[1].score, 2.0)
-            self.assertEqual(updated_trends[2].rank, 3) # Rank should still be distinct
-            self.assertEqual(updated_trends[2].score, 2.0)
-
-            # Check that 'oldpost' tag is not in trends
-            for trend in updated_trends:
-                self.assertNotEqual(trend.hashtag, "oldpost")
-                self.assertNotEqual(trend.hashtag, "old") # from Post 5
-
-
-    def test_get_trending_hashtags_api(self):
-        with self.app.app_context():
-            TrendingHashtag.query.delete()
-            db.session.commit()
-
-            # Manually create TrendingHashtag entries
-            th1 = TrendingHashtag(hashtag="api", score=10.0, rank=1, calculated_at=datetime.utcnow())
-            th2 = TrendingHashtag(hashtag="flask", score=8.0, rank=2, calculated_at=datetime.utcnow())
-            db.session.add_all([th1, th2])
-            db.session.commit()
-
-            response = self.client.get('/api/trending_hashtags')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-
-            self.assertIn('trending_hashtags', data)
-            self.assertEqual(len(data['trending_hashtags']), 2)
-            self.assertEqual(data['trending_hashtags'][0]['hashtag'], "api")
-            self.assertEqual(data['trending_hashtags'][0]['rank'], 1)
-            self.assertEqual(data['trending_hashtags'][1]['hashtag'], "flask")
-            self.assertEqual(data['trending_hashtags'][1]['rank'], 2)
-
-    def test_get_trending_hashtags_api_empty(self):
-        with self.app.app_context():
-            TrendingHashtag.query.delete()
-            db.session.commit()
-
-            response = self.client.get('/api/trending_hashtags')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-
-            self.assertIn('trending_hashtags', data)
-            self.assertEqual(len(data['trending_hashtags']), 0)
-
-
-class TestUserFeedAPI(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # self.user1, self.user2, self.user3 are created by AppTestCase's _setup_base_users()
-
-    def _create_db_like(self, user_id, post_id, timestamp=None):
-        # Duplicating from TestRecommendationAPI, consider moving to AppTestCase if used more widely
-        from models import Like
-        like = Like(user_id=user_id, post_id=post_id, timestamp=timestamp or datetime.utcnow())
-        db.session.add(like)
-        db.session.commit()
-        return like
-
-    def test_get_user_feed_successful_and_structure(self):
-        """ Test Case 1: Successful Feed Retrieval with correct structure, including recommendation reason. """
-        with app.app_context():
-            # user1 is the target, user2 is a friend who makes a post.
-            self._create_friendship(self.user1_id, self.user2_id)
-            post_by_friend = self._create_db_post(user_id=self.user2_id, title="Friend's Post for Feed", content="Content here")
-
-            response = self.client.get(f'/api/users/{self.user1_id}/feed')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.content_type, 'application/json')
-
-            data = response.get_json()
-            self.assertIn('feed_posts', data)
-            feed_posts = data['feed_posts']
-            self.assertIsInstance(feed_posts, list)
-
-            if feed_posts: # Check structure if posts are returned
-                found_friend_post = False
-                for post_data in feed_posts:
-                    self.assertIn('id', post_data)
-                    self.assertIn('title', post_data)
-                    self.assertIn('content', post_data)
-                    self.assertIn('author_username', post_data)
-                    self.assertIn('timestamp', post_data)
-                    # Assert recommendation_reason is present
-                    self.assertIn('recommendation_reason', post_data)
-                    self.assertIsInstance(post_data['recommendation_reason'], str)
-                    # self.assertTrue(len(post_data['recommendation_reason'].strip()) > 0) # Check if non-empty, if that's a strict rule
-
-                    if post_data['id'] == post_by_friend.id:
-                        found_friend_post = True
-                        self.assertEqual(post_data['title'], "Friend's Post for Feed")
-                self.assertTrue(found_friend_post, "Friend's post not found in the feed where it was expected.")
-            else:
-                app.logger.warning("test_get_user_feed_successful_and_structure received an empty feed, which might be unexpected for this test's simple data setup.")
-
-    def test_get_personalized_feed_with_reasons(self):
-        """ Test the personalized feed API specifically for recommendation reasons content. """
-        with app.app_context():
-            # Setup: user1 and user2 are friends. user2 creates a post.
-            self._create_friendship(self.user1_id, self.user2_id, status='accepted')
-            post_by_user2 = self._create_db_post(user_id=self.user2_id, title="Reasonable Post", content="Content that needs a reason.")
-
-            # Execution: Get user1's feed
-            response = self.client.get(f'/api/users/{self.user1_id}/feed')
-
-            # Assertions
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.content_type, 'application/json')
-
-            data = response.get_json()
-            self.assertIn('feed_posts', data)
-            feed_posts = data['feed_posts']
-            self.assertIsInstance(feed_posts, list)
-
-            found_post_with_reason = False
-            if not feed_posts:
-                app.logger.warning(f"Personalized feed for user {self.user1_id} was empty in test_get_personalized_feed_with_reasons. Post by user {self.user2_id} (ID: {post_by_user2.id}) was expected.")
-                # Depending on how critical this is for the test, you might fail here:
-                # self.fail("Personalized feed was empty, cannot verify reason structure.")
-
-            for post_data in feed_posts:
-                self.assertIsInstance(post_data, dict)
-                self.assertIn('id', post_data)
-                self.assertIn('title', post_data)
-                # ... other standard field checks
-
-                self.assertIn('recommendation_reason', post_data)
-                self.assertIsInstance(post_data['recommendation_reason'], str)
-                self.assertTrue(len(post_data['recommendation_reason'].strip()) > 0,
-                                f"Recommendation reason for post ID {post_data.get('id')} was empty or whitespace.")
-
-                if post_data['id'] == post_by_user2.id:
-                    found_post_with_reason = True
-                    # Example check for a specific part of the reason, if predictable:
-                    # self.assertIn("From user you follow", post_data['recommendation_reason'])
-
-            if feed_posts: # Only run if feed_posts is not empty
-                self.assertTrue(found_post_with_reason, f"The specific post (ID: {post_by_user2.id}) by user2 was not found in user1's feed, or was missing a valid reason.")
-
-
-    def test_feed_personalization_friend_vs_non_friend_post(self):
-        """ Test Case 2: Feed personalization (friend's post vs non-friend's post). """
-        with app.app_context():
-            user1 = self.user1
-            user2 = self.user2 # Friend
-            user3 = self.user3 # Non-Friend
-
-            post_from_friend = self._create_db_post(user_id=user2.id, title="Post from Friend")
-            post_from_non_friend = self._create_db_post(user_id=user3.id, title="Post from Non-Friend")
-
-            self._create_friendship(user1.id, user2.id)
-
-            response = self.client.get(f'/api/users/{user1.id}/feed')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            returned_post_ids = {post['id'] for post in data['feed_posts']}
-
-            self.assertIn(post_from_friend.id, returned_post_ids, "Friend's post should be in the feed.")
-            # Assuming the default recommendation logic prioritizes connected content heavily over general content for non-friends
-            # or that non-friend content only appears if it's highly trending.
-            # For this test, we assume post_from_non_friend won't appear without other signals.
-            self.assertNotIn(post_from_non_friend.id, returned_post_ids, "Non-friend's post should not be in the feed without strong other signals.")
-
-    def test_get_feed_user_not_found(self):
-        """ Test Case 3: User Not Found. """
-        with app.app_context():
-            response = self.client.get('/api/users/99999/feed') # Non-existent user ID
-            self.assertEqual(response.status_code, 404)
-            # Flask-RESTful's default 404 for get_or_404 doesn't usually have a JSON body
-            # but if it does, or if a custom error handler is in place:
-            # data = response.get_json()
-            # self.assertIn('message', data)
-
-    @patch('app.api.get_personalized_feed_posts') # Mock the direct source of feed items
-    def test_get_feed_empty_for_new_user_no_relevant_content(self, mock_get_feed_posts_func):
-        """ Test Case 4: Empty Feed for New User (or user with no relevant activity/content). """
-        with app.app_context():
-            mock_get_feed_posts_func.return_value = [] # Ensure recommender returns nothing
-
-            new_user = User(username='newbie', email='newbie@example.com', password_hash=generate_password_hash('password'))
-            db.session.add(new_user)
-            db.session.commit()
-
-            response = self.client.get(f'/api/users/{new_user.id}/feed')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertEqual(data, {'feed_posts': []})
-            mock_get_feed_posts_func.assert_called_once_with(new_user.id, limit=20)
-
-
-    def test_feed_excludes_own_posts(self):
-        """ Test that a user's own posts are not in their feed. """
-        with app.app_context():
-            user1 = self.user1
-            own_post = self._create_db_post(user_id=user1.id, title="My Own Post")
-
-            # To ensure the feed wouldn't be empty otherwise, make user1 friends with user2, and user2 posts.
-            self._create_friendship(user1.id, self.user2_id)
-            self._create_db_post(user_id=self.user2_id, title="Friend's Post to ensure feed is not empty")
-
-
-            response = self.client.get(f'/api/users/{user1.id}/feed')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            returned_post_ids = {post['id'] for post in data['feed_posts']}
-
-            self.assertNotIn(own_post.id, returned_post_ids, "User's own post should not be in their feed.")
-
-    def test_feed_excludes_interacted_posts(self):
-        """ Test that posts liked by the user are not in their feed. """
-        with app.app_context():
-            user1 = self.user1
-            user2 = self.user2 # Friend
-
-            self._create_friendship(user1.id, user2.id)
-
-            post_by_user2_to_be_liked = self._create_db_post(user_id=user2.id, title="Post to be Liked")
-
-            # User1 likes this post
-            self._create_db_like(user_id=user1.id, post_id=post_by_user2_to_be_liked.id)
-
-            # Another post by user2, not interacted with, to ensure feed isn't empty
-            post_by_user2_not_interacted = self._create_db_post(user_id=user2.id, title="Another Post by Friend")
-
-
-            response = self.client.get(f'/api/users/{user1.id}/feed')
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            returned_post_ids = {post['id'] for post in data['feed_posts']}
-
-            self.assertNotIn(post_by_user2_to_be_liked.id, returned_post_ids, "Post liked by user should not be in their feed.")
-            self.assertIn(post_by_user2_not_interacted.id, returned_post_ids, "A non-interacted friend's post should appear for this test.")
-
-
-if __name__ == '__main__':
-    with app.app_context(): # Ensure app context for initial db.create_all() if run directly
-        db.create_all()
-    unittest.main()
-    with app.app_context(): # Ensure app context for final db.drop_all()
-        db.drop_all()
-
-
-class TestOnThisDayAPI(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.test_user = self.user1 # Use user1 from base setup
-        self.fixed_today = datetime(2023, 10, 26, 12, 0, 0) # Fixed date for testing - Oct 26
-
-        # Posts for self.test_user (user1)
-        self.post_target_correct = self._create_db_post(
-            user_id=self.test_user.id, title="Correct Post", content="Content from Oct 26, 2022",
-            timestamp=datetime(2022, 10, 26, 10, 0, 0) # Correct: Same day/month, past year
-        )
-        self.post_current_year = self._create_db_post(
-            user_id=self.test_user.id, title="Current Year Post", content="Content from Oct 26, 2023",
-            timestamp=datetime(2023, 10, 26, 11, 0, 0) # Incorrect: Same day/month, current year
-        )
-        self.post_different_day = self._create_db_post(
-            user_id=self.test_user.id, title="Different Day Post", content="Content from Oct 27, 2022",
-            timestamp=datetime(2022, 10, 27, 12, 0, 0) # Incorrect: Different day
-        )
-        self.post_different_month = self._create_db_post(
-            user_id=self.test_user.id, title="Different Month Post", content="Content from Nov 26, 2022",
-            timestamp=datetime(2022, 11, 26, 12, 0, 0) # Incorrect: Different month
-        )
-        self.post_by_other_user_correct_date = self._create_db_post(
-            user_id=self.user2.id, title="Other User Correct Date Post", content="Content from Oct 26, 2022 by other user",
-            timestamp=datetime(2022, 10, 26, 10, 0, 0) # Correct date, but wrong user
-        )
-
-        # Events for self.test_user (user1)
-        self.event_target_correct = self._create_db_event(
-            user_id=self.test_user.id, title="Correct Event", date_str='2022-10-26',
-            description="Event on Oct 26, 2022" # Correct: Same day/month, past year
-        )
-        self.event_current_year = self._create_db_event(
-            user_id=self.test_user.id, title="Current Year Event", date_str='2023-10-26',
-            description="Event on Oct 26, 2023" # Incorrect: Same day/month, current year
-        )
-        self.event_different_day = self._create_db_event(
-            user_id=self.test_user.id, title="Different Day Event", date_str='2022-10-27',
-            description="Event on Oct 27, 2022" # Incorrect: Different day
-        )
-        self.event_different_month = self._create_db_event(
-            user_id=self.test_user.id, title="Different Month Event", date_str='2022-11-26',
-            description="Event on Nov 26, 2022" # Incorrect: Different month
-        )
-        self.event_by_other_user_correct_date = self._create_db_event(
-            user_id=self.user2.id, title="Other User Correct Date Event", date_str='2022-10-26',
-            description="Event on Oct 26, 2022 by other user" # Correct date, but wrong user
-        )
-        # Event with invalid date format (should be skipped by logic)
-        self.event_invalid_date_format = self._create_db_event(
-            user_id=self.test_user.id, title="Invalid Date Format Event", date_str='2022/10/26' # Wrong format
-        )
-        db.session.commit()
-
-
-    def _create_db_event(self, user_id, title, date_str, description="Test Event Description", time="12:00", location="Test Location", created_at=None):
-        event = Event(
-            user_id=user_id, title=title, description=description,
-            date=date_str, time=time, location=location,
-            created_at=created_at or datetime.utcnow()
-        )
-        db.session.add(event)
-        # db.session.commit() # Commit is handled in setUp after all creations
-        return event
-
-    @patch('recommendations.datetime') # Patch datetime used in recommendations.py
-    @patch('api.datetime') # Patch datetime used in api.py (if any, for current_year logic if not passed from recommendations)
-    def test_on_this_day_with_content_and_filtering(self, mock_api_datetime, mock_reco_datetime):
-        with app.app_context():
-            mock_reco_datetime.utcnow.return_value = self.fixed_today
-            mock_reco_datetime.strptime = datetime.strptime # Ensure strptime is not mocked away for Event date parsing
-            mock_api_datetime.utcnow.return_value = self.fixed_today # If api directly uses it
-
-            token = self._get_jwt_token(self.test_user.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.get('/api/onthisday', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-
-            self.assertIn('on_this_day_posts', data)
-            self.assertIn('on_this_day_events', data)
-
-            # Verify correct post is present
-            self.assertEqual(len(data['on_this_day_posts']), 1, f"Expected 1 post, got {len(data['on_this_day_posts'])}: {data['on_this_day_posts']}")
-            self.assertEqual(data['on_this_day_posts'][0]['id'], self.post_target_correct.id)
-            self.assertEqual(data['on_this_day_posts'][0]['title'], self.post_target_correct.title)
-
-            # Verify correct event is present
-            self.assertEqual(len(data['on_this_day_events']), 1, f"Expected 1 event, got {len(data['on_this_day_events'])}: {data['on_this_day_events']}")
-            self.assertEqual(data['on_this_day_events'][0]['id'], self.event_target_correct.id)
-            self.assertEqual(data['on_this_day_events'][0]['title'], self.event_target_correct.title)
-
-            # Check that other posts/events are NOT present
-            post_ids_in_response = {p['id'] for p in data['on_this_day_posts']}
-            self.assertNotIn(self.post_current_year.id, post_ids_in_response)
-            self.assertNotIn(self.post_different_day.id, post_ids_in_response)
-            self.assertNotIn(self.post_different_month.id, post_ids_in_response)
-            self.assertNotIn(self.post_by_other_user_correct_date.id, post_ids_in_response)
-
-            event_ids_in_response = {e['id'] for e in data['on_this_day_events']}
-            self.assertNotIn(self.event_current_year.id, event_ids_in_response)
-            self.assertNotIn(self.event_different_day.id, event_ids_in_response)
-            self.assertNotIn(self.event_different_month.id, event_ids_in_response)
-            self.assertNotIn(self.event_by_other_user_correct_date.id, event_ids_in_response)
-            self.assertNotIn(self.event_invalid_date_format.id, event_ids_in_response)
-
-    @patch('recommendations.datetime')
-    @patch('api.datetime')
-    def test_on_this_day_no_content(self, mock_api_datetime, mock_reco_datetime):
-        with app.app_context():
-            # Mock current time to a date where no "on this day" content was created for self.test_user
-            no_content_date = datetime(2023, 1, 1, 12, 0, 0)
-            mock_reco_datetime.utcnow.return_value = no_content_date
-            mock_reco_datetime.strptime = datetime.strptime
-            mock_api_datetime.utcnow.return_value = no_content_date
-
-            token = self._get_jwt_token(self.test_user.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.get('/api/onthisday', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-
-            self.assertIn('on_this_day_posts', data)
-            self.assertIn('on_this_day_events', data)
-
-            self.assertEqual(len(data['on_this_day_posts']), 1)
-            self.assertEqual(data['on_this_day_posts'][0]['id'], self.post_on_this_day_past.id)
-            self.assertEqual(data['on_this_day_posts'][0]['title'], self.post_on_this_day_past.title)
-
-            self.assertEqual(len(data['on_this_day_events']), 1)
-            self.assertEqual(data['on_this_day_events'][0]['id'], self.event_on_this_day_past.id)
-            self.assertEqual(data['on_this_day_events'][0]['title'], self.event_on_this_day_past.title)
-
-    @patch('api.datetime')
-    def test_on_this_day_no_content(self, mock_datetime):
-        with app.app_context():
-            # Mock current time to a date where no "on this day" content was created for self.test_user
-            mock_datetime.utcnow.return_value = datetime(2023, 1, 1, 12, 0, 0)
-
-            token = self._get_jwt_token(self.test_user.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.get('/api/onthisday', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = json.loads(response.data)
-
-            self.assertIn('on_this_day_posts', data)
-            self.assertIn('on_this_day_events', data)
-            self.assertEqual(len(data['on_this_day_posts']), 0)
-            self.assertEqual(len(data['on_this_day_events']), 0)
-
-    def test_on_this_day_unauthenticated(self):
-        with app.app_context():
-            response = self.client.get('/api/onthisday')
-            self.assertEqual(response.status_code, 401) # Expecting 401 for missing JWT
-            data = json.loads(response.data)
-            self.assertIn('msg', data) # flask-jwt-extended default error key
-            self.assertEqual(data['msg'], 'Missing Authorization Header')
-
-
-class TestUserStatsAPI(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # user1, user2, user3 are created by AppTestCase's _setup_base_users()
-        # For this test, we'll use self.user1 as the primary user whose stats are fetched.
-        # We need to explicitly set created_at for self.user1 for predictable join_date.
-        with app.app_context():
-            user_to_update = User.query.get(self.user1_id)
-            if user_to_update:
-                user_to_update.created_at = datetime(2023, 1, 1, 12, 0, 0)
-                db.session.add(user_to_update)
-                db.session.commit()
-                self.user1 = User.query.get(self.user1_id) # Re-fetch to ensure session has updated object
-            else:
-                self.fail(f"User with ID {self.user1_id} not found during setUp for TestUserStatsAPI.")
-
-
-    def test_user_stats_api(self):
-        with app.app_context():
-            # Setup data for self.user1
-            # 1. Posts by self.user1 (2 posts)
-            post1_u1 = self._create_db_post(user_id=self.user1.id, title="User1 Post 1")
-            post2_u1 = self._create_db_post(user_id=self.user1.id, title="User1 Post 2")
-
-            # 2. Comments by self.user1 (3 comments)
-            # Assuming self.user1 comments on their own post for simplicity
-            self._create_db_comment(user_id=self.user1.id, post_id=post1_u1.id, content="User1 Comment 1")
-            self._create_db_comment(user_id=self.user1.id, post_id=post1_u1.id, content="User1 Comment 2")
-            self._create_db_comment(user_id=self.user1.id, post_id=post2_u1.id, content="User1 Comment 3")
-
-            # 3. Likes received on self.user1's posts (2 likes received)
-            # self.user2 likes post1_u1
-            self._create_db_like(user_id=self.user2.id, post_id=post1_u1.id)
-            # self.user3 likes post2_u1
-            self._create_db_like(user_id=self.user3.id, post_id=post2_u1.id)
-
-            # 4. Friendships for self.user1 (1 friend)
-            # self.user1 is friends with self.user2
-            self._create_friendship(user1_id=self.user1.id, user2_id=self.user2.id, status='accepted')
-
-            # Log in as self.user1
-            token_user1 = self._get_jwt_token(self.user1.username, 'password')
-            headers_user1 = {'Authorization': f'Bearer {token_user1}'}
-
-            # Make GET request to self.user1's stats endpoint
-            response = self.client.get(f'/api/users/{self.user1.id}/stats', headers=headers_user1)
-
-            # Assertions for successful request
-            self.assertEqual(response.status_code, 200)
-            stats_data = json.loads(response.data)
-
-            self.assertEqual(stats_data['posts_count'], 2)
-            self.assertEqual(stats_data['comments_count'], 3)
-            self.assertEqual(stats_data['likes_received_count'], 2)
-            self.assertEqual(stats_data['friends_count'], 1)
-            self.assertIsNotNone(stats_data['join_date'])
-
-            # Ensure self.user1.created_at is not None before calling isoformat()
-            self.assertIsNotNone(self.user1.created_at, "User1's created_at was not set in setUp")
-            expected_join_date = self.user1.created_at.isoformat()
-            self.assertEqual(stats_data['join_date'], expected_join_date)
-
-            # Test unauthorized access (no token)
-            response_no_token = self.client.get(f'/api/users/{self.user1.id}/stats')
-            self.assertEqual(response_no_token.status_code, 401)
-            data_no_token = json.loads(response_no_token.data)
-            self.assertEqual(data_no_token.get('msg'), 'Missing Authorization Header')
-
-
-            # Test forbidden access (self.user2 tries to get self.user1's stats)
-            token_user2 = self._get_jwt_token(self.user2.username, 'password')
-            headers_user2 = {'Authorization': f'Bearer {token_user2}'}
-            response_forbidden = self.client.get(f'/api/users/{self.user1.id}/stats', headers=headers_user2)
-            self.assertEqual(response_forbidden.status_code, 403)
-            data_forbidden = json.loads(response_forbidden.data)
-            self.assertEqual(data_forbidden.get('message'), 'Unauthorized to view these stats')
-
-
-class TestUserStatus(AppTestCase):
-
-    def test_set_status_full(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            status_text = "Feeling great today!"
-            emoji = "🎉"
-
-            response = self.client.post('/set_status', data={
-                'status_text': status_text,
-                'emoji': emoji
-            }, follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200) # Should redirect to profile
-            self.assertIn(f"/{self.user1.username}", response.request.path) # Check redirected URL path
-
-            # Check flash message
-            self.assertIn("Your status has been updated!", response.get_data(as_text=True))
-
-            # Verify database record
-            user_status = UserStatus.query.filter_by(user_id=self.user1.id).order_by(UserStatus.timestamp.desc()).first()
-            self.assertIsNotNone(user_status)
-            self.assertEqual(user_status.status_text, status_text)
-            self.assertEqual(user_status.emoji, emoji)
-            self.logout()
-
-    def test_set_status_only_text(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            status_text = "Just text, no emoji."
-
-            response = self.client.post('/set_status', data={
-                'status_text': status_text,
-                'emoji': ''
-            }, follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(f"/{self.user1.username}", response.request.path)
-            self.assertIn("Your status has been updated!", response.get_data(as_text=True))
-
-            user_status = UserStatus.query.filter_by(user_id=self.user1.id).order_by(UserStatus.timestamp.desc()).first()
-            self.assertIsNotNone(user_status)
-            self.assertEqual(user_status.status_text, status_text)
-            self.assertIsNone(user_status.emoji) # Should be None if empty string was sent
-            self.logout()
-
-    def test_set_status_only_emoji(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            emoji = "🚀"
-
-            response = self.client.post('/set_status', data={
-                'status_text': '',
-                'emoji': emoji
-            }, follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(f"/{self.user1.username}", response.request.path)
-            self.assertIn("Your status has been updated!", response.get_data(as_text=True))
-
-            user_status = UserStatus.query.filter_by(user_id=self.user1.id).order_by(UserStatus.timestamp.desc()).first()
-            self.assertIsNotNone(user_status)
-            self.assertIsNone(user_status.status_text)
-            self.assertEqual(user_status.emoji, emoji)
-            self.logout()
-
-    def test_set_status_empty_input(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            initial_status_count = UserStatus.query.filter_by(user_id=self.user1.id).count()
-
-            response = self.client.post('/set_status', data={
-                'status_text': '',
-                'emoji': ''
-            }, follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200) # Still redirects to profile
-            self.assertIn(f"/{self.user1.username}", response.request.path)
-
-            # Check flash error message
-            self.assertIn("Status text or emoji must be provided.", response.get_data(as_text=True))
-
-            # Verify no new status record was created
-            final_status_count = UserStatus.query.filter_by(user_id=self.user1.id).count()
-            self.assertEqual(final_status_count, initial_status_count)
-            self.logout()
-
-    def test_view_status_on_profile(self):
-        with app.app_context():
-            # Directly create a status for user1
-            status_text = "Testing profile view."
-            emoji = "👀"
-            timestamp = datetime.utcnow() - timedelta(minutes=5)
-            UserStatus.query.delete() # Clear any existing statuses for this user for predictability
-            db.session.commit()
-
-            created_status = UserStatus(user_id=self.user1.id, status_text=status_text, emoji=emoji, timestamp=timestamp)
-            db.session.add(created_status)
-            db.session.commit()
-
-            # Log in as user2 (or user1, doesn't matter for viewing)
-            self.login(self.user2.username, 'password')
-            response = self.client.get(f'/user/{self.user1.username}')
-
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            self.assertIn(status_text, response_data)
-            self.assertIn(emoji, response_data)
-            # Check for formatted timestamp (e.g., minutes, not seconds for robustness if formatting is general)
-            self.assertIn(timestamp.strftime('%Y-%m-%d %H:%M'), response_data)
-            self.logout()
-
-    def test_view_status_on_profile_no_status(self):
-        with app.app_context():
-            # Ensure user1 has no statuses
-            UserStatus.query.filter_by(user_id=self.user1.id).delete()
-            db.session.commit()
-
-            self.login(self.user2.username, 'password') # Login as another user
-            response = self.client.get(f'/user/{self.user1.username}')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            # Assert that status-related HTML elements are NOT present or specific message is shown
-            # Depending on template, it might just be empty or have a placeholder.
-            # For now, let's ensure the status specific classes/text aren't there.
-            self.assertNotIn("user-status", response_data) # Assuming a wrapper div class
-            self.assertNotIn("Status set on:", response_data)
-            self.logout()
-
-
-    def test_set_status_form_visible_on_own_profile(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            response = self.client.get(f'/user/{self.user1.username}')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            # Check for form elements
-            self.assertIn('action="/set_status"', response_data) # More specific than url_for
-            self.assertIn('name="status_text"', response_data)
-            self.assertIn('name="emoji"', response_data)
-            self.assertIn('type="submit"', response_data)
-            self.assertIn("Set Status", response_data) # Button text
-            self.logout()
-
-    def test_set_status_form_not_visible_on_others_profile(self):
-        with app.app_context():
-            # Log in as user1
-            self.login(self.user1.username, 'password')
-
-            # View user2's profile
-            response = self.client.get(f'/user/{self.user2.username}')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            # Assert form elements are NOT present
-            self.assertNotIn('action="/set_status"', response_data)
-            self.assertNotIn('name="status_text"', response_data)
-            self.assertNotIn('name="emoji"', response_data)
-            # Check that the specific "Set Status" button text is not there,
-            # being careful if other forms might exist.
-            # A more robust check might be to ensure the specific form container isn't there.
-            # For now, checking for the action and input names is a good start.
-            self.logout()
-
-
-# Helper to create UserActivity for tests
-def _create_db_user_activity(user_id, activity_type, related_id=None, target_user_id=None, content_preview=None, link=None, timestamp=None):
-    from models import UserActivity # Local import if needed or ensure it's available
-    activity = UserActivity(
-        user_id=user_id,
-        activity_type=activity_type,
-        related_id=related_id,
-        target_user_id=target_user_id,
-        content_preview=content_preview,
-        link=link,
-        timestamp=timestamp or datetime.utcnow()
-    )
-    db.session.add(activity)
-    db.session.commit()
-    return activity
-
-class TestLiveActivityFeed(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # self.user1, self.user2, self.user3 are created by AppTestCase
-        # Make user1 and user3 friends with user2 for socketio emit tests
-        self._create_friendship(self.user2_id, self.user1_id, status='accepted') # user2 is friends with user1
-        self._create_friendship(self.user2_id, self.user3_id, status='accepted') # user2 is friends with user3
-
-    @patch('app.socketio.emit')
-    def test_new_follow_activity_logging_and_socketio(self, mock_socketio_emit):
-        with app.app_context():
-            # user1 sends friend request to user2
-            friend_req = Friendship(user_id=self.user1_id, friend_id=self.user2_id, status='pending')
-            db.session.add(friend_req)
-            db.session.commit()
-            request_id = friend_req.id
-
-            # user2 logs in and accepts the request
-            self.login(self.user2.username, 'password')
-            response = self.client.post(f'/friend_request/{request_id}/accept', follow_redirects=True)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("Friend request accepted successfully!", response.get_data(as_text=True))
-
-            activity = UserActivity.query.filter_by(user_id=self.user2_id, activity_type='new_follow').first()
-            self.assertIsNotNone(activity)
-            self.assertEqual(activity.target_user_id, self.user1_id)
-            self.assertIn(self.user1.username, activity.link) # Link should be to the followed user's profile
-
-            # Check socketio emit (user2 accepted, so event goes to friends of user2)
-            # user2 is friends with user1 and user3 from setUp.
-            # Activity is user2 following user1.
-            # So, user2's friends (user1, user3) should get the event.
-            # However, user1 is the target_user, so the event might be less relevant to them in some designs,
-            # but the current emit_new_activity_event sends to all friends of the actor.
-
-            # Expected payload for emit_new_activity_event
-            # actor is user2, target_user is user1
-            expected_payload = {
-                'activity_id': activity.id,
-                'user_id': self.user2_id,
-                'username': self.user2.username,
-                'profile_picture': self.user2.profile_picture if self.user2.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_follow',
-                'related_id': None, # No related_id for 'new_follow' typically
-                'content_preview': None, # No content_preview for 'new_follow'
-                'link': activity.link,
-                'timestamp': ANY, # Handled by ANY for simplicity
-                'target_user_id': self.user1_id,
-                'target_username': self.user1.username,
-            }
-
-            # Check call for user1 (friend of user2)
-            mock_socketio_emit.assert_any_call('new_activity_event', expected_payload, room=f'user_{self.user1_id}')
-            # Check call for user3 (friend of user2)
-            mock_socketio_emit.assert_any_call('new_activity_event', expected_payload, room=f'user_{self.user3_id}')
-
-            # Ensure it wasn't called for user2 (actor)
-            # This requires checking all calls if emit was called multiple times for other reasons.
-            # For simplicity, we'll trust emit_new_activity_event's internal logic for not sending to self.
-            # A more rigorous check:
-            called_rooms = [call_args[1].get('room') for call_args in mock_socketio_emit.call_args_list if call_args[0][0] == 'new_activity_event']
-            self.assertNotIn(f'user_{self.user2_id}', called_rooms)
-
-            self.logout()
-
-    def test_live_feed_unauthorized_access(self):
-        with app.app_context():
-            response = self.client.get('/live_feed', follow_redirects=False)
-            self.assertEqual(response.status_code, 302) # Redirects to login
-            self.assertIn('/login', response.location)
-
-    def test_live_feed_authorized_access_and_data(self):
-        with app.app_context():
-            # User main (user1), friend1 (user2), friend2 (user3)
-            # non_friend (create a new one)
-            user_non_friend = User(username='nonfriend', email='nf@example.com', password_hash=generate_password_hash('password'))
-            db.session.add(user_non_friend)
-            db.session.commit()
-
-            # Friendships: user1 is friends with user2 and user3
-            self._create_friendship(self.user1_id, self.user2_id)
-            self._create_friendship(self.user1_id, self.user3_id)
-
-            # Activities by friends
-            act_friend1_post = _create_db_user_activity(user_id=self.user2_id, activity_type='new_post', related_id=1, content_preview="Friend1 post", timestamp=datetime.utcnow() - timedelta(minutes=10))
-            act_friend2_like = _create_db_user_activity(user_id=self.user3_id, activity_type='new_like', related_id=2, content_preview="Friend2 liked something", timestamp=datetime.utcnow() - timedelta(minutes=5))
-
-            # Activity by user_main (should not be in their own live feed)
-            _create_db_user_activity(user_id=self.user1_id, activity_type='new_post', related_id=3, content_preview="My own post", timestamp=datetime.utcnow() - timedelta(minutes=1))
-            # Activity by non_friend (should not be in user1's live feed)
-            _create_db_user_activity(user_id=user_non_friend.id, activity_type='new_post', related_id=4, content_preview="Non-friend post", timestamp=datetime.utcnow() - timedelta(minutes=2))
-
-
-            self.login(self.user1.username, 'password')
-            response = self.client.get('/live_feed')
-            self.assertEqual(response.status_code, 200)
-            self.assert_template_used('live_feed.html') # Requires Flask-Testing or custom setup for this assertion
-
-            response_data = response.get_data(as_text=True)
-            self.assertIn(self.user2.username, response_data) # friend1's activity
-            self.assertIn("Friend1 post", response_data)
-            self.assertIn(self.user3.username, response_data) # friend2's activity
-            self.assertIn("Friend2 liked something", response_data)
-
-            self.assertNotIn("My own post", response_data)
-            self.assertNotIn(user_non_friend.username, response_data)
-            self.assertNotIn("Non-friend post", response_data)
-
-            # Check order (friend2's like is newer)
-            self.assertTrue(response_data.find("Friend2 liked something") < response_data.find("Friend1 post"))
-
-            self.logout()
-
-    @patch('app.socketio.emit')
-    def test_emit_new_activity_event_helper_direct(self, mock_socketio_emit):
-        from app import emit_new_activity_event # Import the helper
-        with app.app_context():
-            # user2 is the actor, friends with user1 and user3 (setup in TestLiveActivityFeed.setUp)
-            actor = self.user2
-
-            # Create a sample UserActivity object manually for actor (user2)
-            # This activity is by user2, about a post they made (hypothetically)
-            activity = UserActivity(
-                id=100, # Assign a mock ID
-                user_id=actor.id,
-                user=actor, # Link the user object
-                activity_type='new_post',
-                related_id=50,
-                content_preview='Helper test post preview',
-                link='/blog/post/50',
-                timestamp=datetime.utcnow()
-            )
-            # No need to add to DB for this direct test if User.get_friends() doesn't rely on DB state for these specific users
-
-            emit_new_activity_event(activity)
-
-            expected_payload = {
-                'activity_id': 100,
-                'user_id': actor.id,
-                'username': actor.username,
-                'profile_picture': actor.profile_picture if actor.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_post',
-                'related_id': 50,
-                'content_preview': 'Helper test post preview',
-                'link': '/blog/post/50',
-                'timestamp': activity.timestamp.isoformat(),
-                'target_user_id': None,
-                'target_username': None,
-            }
-
-            # user2 is friends with user1 and user3
-            calls = [
-                call('new_activity_event', expected_payload, room=f'user_{self.user1_id}'),
-                call('new_activity_event', expected_payload, room=f'user_{self.user3_id}')
-            ]
-            mock_socketio_emit.assert_has_calls(calls, any_order=True)
-            self.assertEqual(mock_socketio_emit.call_count, 2) # Assuming user2 only has user1 and user3 as friends in this test context.
-
-            # Test with a 'new_follow' activity
-            mock_socketio_emit.reset_mock()
-
-            followed_user = self.user1 # user2 follows user1
-            follow_activity = UserActivity(
-                id=101,
-                user_id=actor.id,
-                user=actor,
-                activity_type='new_follow',
-                target_user_id=followed_user.id,
-                target_user=followed_user, # Link the target_user object
-                link=f'/user/{followed_user.username}',
-                timestamp=datetime.utcnow()
-            )
-
-            emit_new_activity_event(follow_activity)
-
-            expected_follow_payload = {
-                'activity_id': 101,
-                'user_id': actor.id,
-                'username': actor.username,
-                'profile_picture': actor.profile_picture if actor.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_follow',
-                'related_id': None,
-                'content_preview': None,
-                'link': f'/user/{followed_user.username}',
-                'timestamp': follow_activity.timestamp.isoformat(),
-                'target_user_id': followed_user.id,
-                'target_username': followed_user.username,
-            }
-            calls_follow = [
-                call('new_activity_event', expected_follow_payload, room=f'user_{self.user1_id}'), # user1 is a friend of actor user2
-                call('new_activity_event', expected_follow_payload, room=f'user_{self.user3_id}'), # user3 is a friend of actor user2
-            ]
-            mock_socketio_emit.assert_has_calls(calls_follow, any_order=True)
-            self.assertEqual(mock_socketio_emit.call_count, 2)
-
-    @patch('app.socketio.emit')
-    def test_new_post_activity_logging_and_socketio(self, mock_socketio_emit):
-        with app.app_context():
-            # user2 (actor) creates a post. Friends are user1 and user3.
-            self.login(self.user2.username, 'password')
-            post_title = "User2 New Post for SocketIO Test"
-            post_content = "SocketIO content here."
-            response = self.client.post('/blog/create', data=dict(
-                title=post_title,
-                content=post_content,
-                hashtags="test,socketio"
-            ), follow_redirects=True)
-            self.assertEqual(response.status_code, 200) # Assuming redirect to blog page
-            self.assertIn("Blog post created successfully!", response.get_data(as_text=True))
-
-            created_post = Post.query.filter_by(user_id=self.user2_id, title=post_title).first()
-            self.assertIsNotNone(created_post)
-
-            activity = UserActivity.query.filter_by(user_id=self.user2_id, activity_type='new_post', related_id=created_post.id).first()
-            self.assertIsNotNone(activity)
-            self.assertEqual(activity.content_preview, post_content[:100])
-            self.assertIn(f'/blog/post/{created_post.id}', activity.link)
-
-            expected_payload = {
-                'activity_id': activity.id,
-                'user_id': self.user2_id,
-                'username': self.user2.username,
-                'profile_picture': self.user2.profile_picture if self.user2.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_post',
-                'related_id': created_post.id,
-                'content_preview': post_content[:100],
-                'link': ANY, # Link is generated with url_for, check presence and part of it
-                'timestamp': ANY,
-                'target_user_id': None,
-                'target_username': None,
-            }
-
-            # Check calls for friends of user2 (user1 and user3)
-            calls = [
-                call('new_activity_event', expected_payload, room=f'user_{self.user1_id}'),
-                call('new_activity_event', expected_payload, room=f'user_{self.user3_id}')
-            ]
-            mock_socketio_emit.assert_has_calls(calls, any_order=True)
-            self.assertEqual(mock_socketio_emit.call_count, 2) # Only 2 friends
-            self.logout()
-
-    @patch('app.socketio.emit')
-    def test_new_comment_activity_logging_and_socketio(self, mock_socketio_emit):
-        with app.app_context():
-            # user1 creates a post
-            post_by_user1 = self._create_db_post(user_id=self.user1_id, title="Post to be commented on")
-
-            # user2 (actor) comments on user1's post. Friends of user2 are user1 and user3.
-            self.login(self.user2.username, 'password')
-            comment_content = "A insightful comment by user2."
-            response = self.client.post(f'/blog/post/{post_by_user1.id}/comment', data=dict(
-                comment_content=comment_content
-            ), follow_redirects=True)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("Comment added successfully!", response.get_data(as_text=True))
-
-            # Verify UserActivity
-            activity = UserActivity.query.filter_by(user_id=self.user2_id, activity_type='new_comment', related_id=post_by_user1.id).first()
-            self.assertIsNotNone(activity)
-            self.assertEqual(activity.content_preview, comment_content[:100])
-            self.assertIn(f'/blog/post/{post_by_user1.id}', activity.link)
-
-            expected_payload = {
-                'activity_id': activity.id,
-                'user_id': self.user2_id,
-                'username': self.user2.username,
-                'profile_picture': self.user2.profile_picture if self.user2.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_comment',
-                'related_id': post_by_user1.id,
-                'content_preview': comment_content[:100],
-                'link': ANY,
-                'timestamp': ANY,
-                'target_user_id': None,
-                'target_username': None,
-            }
-            calls = [
-                call('new_activity_event', expected_payload, room=f'user_{self.user1_id}'),
-                call('new_activity_event', expected_payload, room=f'user_{self.user3_id}')
-            ]
-            mock_socketio_emit.assert_has_calls(calls, any_order=True)
-            self.assertEqual(mock_socketio_emit.call_count, 2)
-            self.logout()
-
-    @patch('app.socketio.emit')
-    def test_new_like_activity_logging_and_socketio(self, mock_socketio_emit):
-        with app.app_context():
-            # user1 creates a post
-            post_by_user1 = self._create_db_post(user_id=self.user1_id, title="Post to be liked")
-
-            # user2 (actor) likes user1's post. Friends of user2 are user1 and user3.
-            self.login(self.user2.username, 'password')
-            response = self.client.post(f'/blog/post/{post_by_user1.id}/like', follow_redirects=True)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("Post liked!", response.get_data(as_text=True))
-
-            activity = UserActivity.query.filter_by(user_id=self.user2_id, activity_type='new_like', related_id=post_by_user1.id).first()
-            self.assertIsNotNone(activity)
-            self.assertEqual(activity.content_preview, post_by_user1.content[:100]) # Assuming post content is used for preview
-            self.assertIn(f'/blog/post/{post_by_user1.id}', activity.link)
-
-            expected_payload = {
-                'activity_id': activity.id,
-                'user_id': self.user2_id,
-                'username': self.user2.username,
-                'profile_picture': self.user2.profile_picture if self.user2.profile_picture else '/static/profile_pics/default.png',
-                'activity_type': 'new_like',
-                'related_id': post_by_user1.id,
-                'content_preview': post_by_user1.content[:100],
-                'link': ANY,
-                'timestamp': ANY,
-                'target_user_id': None,
-                'target_username': None,
-            }
-            calls = [
-                call('new_activity_event', expected_payload, room=f'user_{self.user1_id}'),
-                call('new_activity_event', expected_payload, room=f'user_{self.user3_id}')
-            ]
-            mock_socketio_emit.assert_has_calls(calls, any_order=True)
-            self.assertEqual(mock_socketio_emit.call_count, 2)
-            self.logout()
-
-
-class TestFileSharing(AppTestCase):
-
-    def create_dummy_file(self, filename="test.txt", content=b"hello world", content_type="text/plain"):
-        return (io.BytesIO(content), filename, content_type)
-
-    def test_share_file_get_page(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            response = self.client.get(f'/files/share/{self.user2.username}')
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(f"Share File with {self.user2.username}", response.get_data(as_text=True))
-            self.logout()
-
-    def test_share_file_successful_upload(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="upload_test.txt", content=b"Test file content for upload.")
-
-            data = {
-                'file': dummy_file_data,
-                'message': "This is a test message for the shared file."
-            }
-            response = self.client.post(f'/files/share/{self.user2.username}', data=data, content_type='multipart/form-data', follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200) # After redirect, should land on inbox
-            self.assertIn("File successfully shared!", response.get_data(as_text=True))
-
-            shared_file_record = SharedFile.query.filter_by(sender_id=self.user1_id, receiver_id=self.user2_id).first()
-            self.assertIsNotNone(shared_file_record)
-            self.assertEqual(shared_file_record.original_filename, "upload_test.txt")
-            self.assertEqual(shared_file_record.message, "This is a test message for the shared file.")
-
-            file_path = os.path.join(app.config['SHARED_FILES_UPLOAD_FOLDER'], shared_file_record.saved_filename)
-            self.assertTrue(os.path.exists(file_path))
-
-            # Cleanup (file will be cleaned by tearDown, record needs manual delete or rely on tearDown's db.drop_all)
-            # For explicit test cleanup:
-            # os.remove(file_path)
-            # db.session.delete(shared_file_record)
-            # db.session.commit()
-            self.logout()
-
-    def test_share_file_invalid_file_type(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="test.exe", content=b"executable content", content_type="application/octet-stream")
-
-            data = {'file': dummy_file_data}
-            response = self.client.post(f'/files/share/{self.user2.username}', data=data, content_type='multipart/form-data', follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("File type not allowed", response.get_data(as_text=True))
-            self.assertIsNone(SharedFile.query.filter_by(original_filename="test.exe").first())
-            self.logout()
-
-    def test_share_file_too_large(self):
-        with app.app_context():
-            original_max_size = app.config['SHARED_FILES_MAX_SIZE']
-            app.config['SHARED_FILES_MAX_SIZE'] = 10 # 10 bytes
-
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="large_file.txt", content=b"This content is definitely larger than 10 bytes.")
-
-            data = {'file': dummy_file_data}
-            response = self.client.post(f'/files/share/{self.user2.username}', data=data, content_type='multipart/form-data', follow_redirects=True)
-
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("File is too large", response.get_data(as_text=True))
-            self.assertIsNone(SharedFile.query.filter_by(original_filename="large_file.txt").first())
-
-            app.config['SHARED_FILES_MAX_SIZE'] = original_max_size # Restore original config
-            self.logout()
-
-    def test_files_inbox_empty(self):
-        with app.app_context():
-            self.login(self.user2.username, 'password')
-            response = self.client.get('/files/inbox')
-            self.assertEqual(response.status_code, 200)
-            self.assertIn("You have not received any files.", response.get_data(as_text=True))
-            self.logout()
-
-    def test_files_inbox_with_files(self):
-        with app.app_context():
-            # user1 shares file with user2
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="inbox_test_file.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data, 'message': 'Hi!'}, content_type='multipart/form-data')
-            self.logout()
-
-            # user2 logs in and checks inbox
-            self.login(self.user2.username, 'password')
-            response = self.client.get('/files/inbox')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-            self.assertIn("inbox_test_file.txt", response_data)
-            self.assertIn(self.user1.username, response_data) # Sender's username
-            self.assertIn("Hi!", response_data) # Message
-            self.assertIn("(New)", response_data) # Unread indicator
-            self.logout()
-
-    def test_download_shared_file_receiver(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="download_me.txt", content=b"Downloadable content.")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="download_me.txt").first()
-            self.assertIsNotNone(shared_file)
-            self.assertFalse(shared_file.is_read)
-            self.logout()
-
-            self.login(self.user2.username, 'password')
-            response = self.client.get(f'/files/download/{shared_file.id}')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.mimetype, 'text/plain') # Based on .txt, might vary
-            self.assertIn('attachment; filename=download_me.txt', response.headers['Content-Disposition'])
-            self.assertEqual(response.data, b"Downloadable content.")
-
-            db.session.refresh(shared_file) # Refresh from DB
-            self.assertTrue(shared_file.is_read)
-            self.logout()
-
-    def test_download_shared_file_sender(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="sender_download.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="sender_download.txt").first()
-            self.assertIsNotNone(shared_file)
-            initial_is_read_status = shared_file.is_read # Should be False
-            self.logout()
-
-            self.login(self.user1.username, 'password') # Sender logs back in
-            response = self.client.get(f'/files/download/{shared_file.id}')
-            self.assertEqual(response.status_code, 200)
-
-            db.session.refresh(shared_file)
-            self.assertEqual(shared_file.is_read, initial_is_read_status) # is_read should not change for sender
-            self.logout()
-
-    def test_download_shared_file_unauthorized(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="unauth_download.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="unauth_download.txt").first()
-            self.assertIsNotNone(shared_file)
-            self.logout()
-
-            self.login(self.user3.username, 'password') # Unauthorized user
-            response = self.client.get(f'/files/download/{shared_file.id}', follow_redirects=True)
-            self.assertEqual(response.status_code, 200) # After redirect
-            self.assertIn("You are not authorized to download this file.", response.get_data(as_text=True))
-            self.logout()
-
-    def test_delete_shared_file_receiver(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="to_delete_receiver.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="to_delete_receiver.txt").first()
-            self.assertIsNotNone(shared_file)
-            saved_filename = shared_file.saved_filename
-            file_path = os.path.join(app.config['SHARED_FILES_UPLOAD_FOLDER'], saved_filename)
-            self.assertTrue(os.path.exists(file_path))
-            self.logout()
-
-            self.login(self.user2.username, 'password') # Receiver logs in
-            response = self.client.post(f'/files/delete/{shared_file.id}', follow_redirects=True)
-            self.assertEqual(response.status_code, 200) # After redirect to inbox
-            self.assertIn("File successfully deleted.", response.get_data(as_text=True))
-
-            self.assertIsNone(SharedFile.query.get(shared_file.id))
-            self.assertFalse(os.path.exists(file_path)) # File should be gone
-            self.logout()
-
-    def _create_series(self, user_id, title="Test Series", description="A series for testing.", created_at=None, updated_at=None):
-        series = Series(
-            title=title,
-            description=description,
-            user_id=user_id,
-            created_at=created_at or datetime.utcnow(),
-            updated_at=updated_at or datetime.utcnow()
-        )
-        db.session.add(series)
-        db.session.commit()
-        return series
-
-
-# Helper function to seed achievements for tests
+# TestRealtimePostNotifications, TestLikeNotifications, AchievementLogicTests,
+# and seed_test_achievements helper have been moved to their respective files.
 
 class TestMinimalSanityCheck(unittest.TestCase):
     def test_absolutely_nothing(self):
         self.assertTrue(True)
 
-class TestCollaborativeEditing(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # self.user1, self.user2, self.user3 are created by AppTestCase
-        self.post_author = self.user1
-        self.collaborator = self.user2
-        self.another_user = self.user3
-
-        self.test_post = self._create_db_post(user_id=self.post_author.id, title="Collaborative Post", content="Initial content.")
-
-        # Initialize SocketIO test client for this test class
-        self.socketio_client = socketio.test_client(app)
-        # You might want to connect the client here or in each test method that needs it.
-        # self.socketio_client.connect() # Connect if needed globally for tests
-
-    def tearDown(self):
-        # Disconnect the SocketIO test client if connected
-        if self.socketio_client.connected:
-            self.socketio_client.disconnect()
-        super().tearDown()
-
-    def _create_db_lock(self, post_id, user_id, minutes_offset=0):
-        """Helper to create a PostLock directly in the DB."""
-        from models import PostLock # Ensure PostLock is imported
-        expires_at = datetime.utcnow() + timedelta(minutes=minutes_offset)
-        lock = PostLock(post_id=post_id, user_id=user_id, expires_at=expires_at)
-        db.session.add(lock)
-        db.session.commit()
-        return lock
-
-    # --- Model Tests ---
-    def test_post_lock_creation(self):
-        with app.app_context():
-            lock = self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=15)
-            self.assertIsNotNone(lock.id)
-            self.assertEqual(lock.post_id, self.test_post.id)
-            self.assertEqual(lock.user_id, self.post_author.id)
-            self.assertTrue(lock.expires_at > datetime.utcnow())
-
-    def test_post_is_locked_method(self):
-        with app.app_context():
-            # 1. No lock
-            self.assertFalse(self.test_post.is_locked(), "Post should not be locked if no lock exists.")
-
-            # 2. Active lock
-            active_lock = self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=15)
-            db.session.refresh(self.test_post) # Refresh to load lock_info relationship
-            self.assertTrue(self.test_post.is_locked(), "Post should be locked with an active lock.")
-
-            db.session.delete(active_lock) # Clean up for next check
-            db.session.commit()
-            db.session.refresh(self.test_post) # Refresh again
-
-            # 3. Expired lock
-            expired_lock = self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=-5) # Lock expired 5 mins ago
-            db.session.refresh(self.test_post)
-            self.assertFalse(self.test_post.is_locked(), "Post should not be locked if the lock is expired.")
-            db.session.delete(expired_lock)
-            db.session.commit()
-
-    # --- API Endpoint Tests (PostLockResource) ---
-    def test_api_acquire_lock_success(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.collaborator.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.post(f'/api/posts/{self.test_post.id}/lock', headers=headers)
-            self.assertEqual(response.status_code, 200, f"Response data: {response.data.decode()}")
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Post locked successfully.')
-            self.assertEqual(data['lock_details']['post_id'], self.test_post.id)
-            self.assertEqual(data['lock_details']['locked_by_user_id'], self.collaborator.id)
-
-            lock_in_db = PostLock.query.filter_by(post_id=self.test_post.id).first()
-            self.assertIsNotNone(lock_in_db)
-            self.assertEqual(lock_in_db.user_id, self.collaborator.id)
-
-    def test_api_acquire_lock_conflict(self):
-        with app.app_context():
-            # post_author locks the post first
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=15)
-
-            # collaborator tries to lock it
-            token_collab = self._get_jwt_token(self.collaborator.username, 'password')
-            headers_collab = {'Authorization': f'Bearer {token_collab}'}
-            response = self.client.post(f'/api/posts/{self.test_post.id}/lock', headers=headers_collab)
-
-            self.assertEqual(response.status_code, 409) # Conflict
-            data = response.get_json()
-            self.assertIn('Post is currently locked by another user.', data['message'])
-            self.assertEqual(data['locked_by_username'], self.post_author.username)
-
-    def test_api_acquire_own_expired_lock(self):
-        with app.app_context():
-            # collaborator had an old, expired lock
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.collaborator.id, minutes_offset=-5)
-
-            token_collab = self._get_jwt_token(self.collaborator.username, 'password')
-            headers_collab = {'Authorization': f'Bearer {token_collab}'}
-            response = self.client.post(f'/api/posts/{self.test_post.id}/lock', headers=headers_collab)
-
-            self.assertEqual(response.status_code, 200) # Should succeed, replacing old lock
-            data = response.get_json()
-            self.assertEqual(data['lock_details']['locked_by_user_id'], self.collaborator.id)
-
-            locks_in_db = PostLock.query.filter_by(post_id=self.test_post.id).all()
-            self.assertEqual(len(locks_in_db), 1) # Old one deleted, new one created
-            self.assertTrue(locks_in_db[0].expires_at > datetime.utcnow())
-
-    def test_api_acquire_lock_unauthenticated(self):
-        with app.app_context():
-            response = self.client.post(f'/api/posts/{self.test_post.id}/lock') # No token
-            self.assertEqual(response.status_code, 401)
-
-    def test_api_release_lock_success_owner(self):
-        with app.app_context():
-            # collaborator acquires lock
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.collaborator.id, minutes_offset=15)
-
-            token_collab = self._get_jwt_token(self.collaborator.username, 'password')
-            headers_collab = {'Authorization': f'Bearer {token_collab}'}
-            response = self.client.delete(f'/api/posts/{self.test_post.id}/lock', headers=headers_collab)
-
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Post unlocked successfully.')
-            self.assertIsNone(PostLock.query.filter_by(post_id=self.test_post.id).first())
-
-    def test_api_release_lock_not_owner(self):
-        with app.app_context():
-            # post_author has the lock
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=15)
-
-            # collaborator tries to release it
-            token_collab = self._get_jwt_token(self.collaborator.username, 'password')
-            headers_collab = {'Authorization': f'Bearer {token_collab}'}
-            response = self.client.delete(f'/api/posts/{self.test_post.id}/lock', headers=headers_collab)
-
-            self.assertEqual(response.status_code, 403) # Forbidden
-            data = response.get_json()
-            self.assertIn('You are not authorized to unlock this post', data['message'])
-            self.assertIsNotNone(PostLock.query.filter_by(post_id=self.test_post.id).first()) # Lock still there
-
-    def test_api_release_lock_unauthenticated(self):
-        with app.app_context():
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.post_author.id, minutes_offset=15)
-            response = self.client.delete(f'/api/posts/{self.test_post.id}/lock') # No token
-            self.assertEqual(response.status_code, 401)
-            self.assertIsNotNone(PostLock.query.filter_by(post_id=self.test_post.id).first())
-
-    def test_api_release_lock_not_locked(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.collaborator.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.delete(f'/api/posts/{self.test_post.id}/lock', headers=headers)
-            self.assertEqual(response.status_code, 404) # Post not locked
-
-    # --- SocketIO Event Tests ---
-    # Note: Testing SocketIO broadcasts to rooms requires careful handling of the test client.
-    # socketio.test_client(app) allows receiving events emitted globally or to specific SIDs.
-    # For room broadcasts, you might need to have clients join rooms or mock `socketio.emit`.
-
-    @patch('app.socketio.emit') # Mock socketio.emit in the app module where the handler calls it
-    def test_socketio_edit_post_with_lock(self, mock_app_socketio_emit):
-        with app.app_context():
-            # 1. Collaborator acquires lock via API (this also tests API's emit for lock_acquired)
-            token_collab = self._get_jwt_token(self.collaborator.username, 'password')
-            headers_collab = {'Authorization': f'Bearer {token_collab}'}
-            self.client.post(f'/api/posts/{self.test_post.id}/lock', headers=headers_collab)
-
-            # We need to simulate the collaborator being connected via SocketIO
-            # For this test, we'll directly call the handler logic as if from a socket event.
-            # A more integrated test would use self.socketio_client.emit and .get_received()
-
-            # Simulate collaborator being logged into session for the SocketIO handler
-            with self.client.session_transaction() as sess:
-                sess['user_id'] = self.collaborator.id
-                sess['username'] = self.collaborator.username
-
-            # Simulate 'edit_post_content' event from collaborator
-            new_content = "Updated collaboratively!"
-            socketio.on_event('edit_post_content', handler=app.handle_edit_post_content) # Ensure handler is registered if not auto
-
-            # Emit the event using the main socketio instance (as if client sent it)
-            # This requires the test client to be connected or the event to be handled in current context.
-            # For simplicity, let's directly invoke the handler function or use the test_client to emit.
-
-            # Using the test_client to emit and receive
-            client = socketio.test_client(app, flask_test_client=self.client)
-            client.emit('edit_post_content', {'post_id': self.test_post.id, 'new_content': new_content})
-
-            db.session.refresh(self.test_post)
-            self.assertEqual(self.test_post.content, new_content)
-
-            # Check for 'post_content_updated' broadcast
-            # mock_app_socketio_emit should have been called by handle_edit_post_content
-            # We need to find the right call among potentially multiple calls (e.g. from lock acquisition)
-            found_content_update_emit = False
-            for call_args_tuple in mock_app_socketio_emit.call_args_list:
-                event_name = call_args_tuple[0][0]
-                if event_name == 'post_content_updated':
-                    payload = call_args_tuple[0][1]
-                    room = call_args_tuple[1].get('room')
-                    self.assertEqual(payload['post_id'], self.test_post.id)
-                    self.assertEqual(payload['new_content'], new_content)
-                    self.assertEqual(payload['edited_by_user_id'], self.collaborator.id)
-                    self.assertEqual(room, f'post_{self.test_post.id}')
-                    found_content_update_emit = True
-                    break
-            self.assertTrue(found_content_update_emit, "post_content_updated event was not emitted correctly.")
-            client.disconnect()
-
-
-    def test_socketio_edit_post_without_lock(self):
-        with app.app_context():
-            # Ensure no lock exists
-            PostLock.query.delete()
-            db.session.commit()
-
-            # Simulate another_user (who doesn't have a lock) trying to edit
-            with self.client.session_transaction() as sess:
-                sess['user_id'] = self.another_user.id
-
-            client = socketio.test_client(app, flask_test_client=self.client)
-            client.emit('edit_post_content', {'post_id': self.test_post.id, 'new_content': "Sneaky update"})
-
-            received = client.get_received()
-            self.assertTrue(any(e['name'] == 'edit_error' for e in received), "edit_error should have been emitted.")
-            edit_error_event = next(e for e in received if e['name'] == 'edit_error')
-            self.assertIn('Post is not locked', edit_error_event['args'][0]['message'])
-
-            db.session.refresh(self.test_post)
-            self.assertEqual(self.test_post.content, "Initial content.") # Content should not change
-            client.disconnect()
-
-    @patch('api.socketio.emit') # Mock socketio.emit in the api module
-    def test_socketio_lock_acquired_broadcast_from_api(self, mock_api_socketio_emit):
-        with app.app_context():
-            token = self._get_jwt_token(self.collaborator.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.post(f'/api/posts/{self.test_post.id}/lock', headers=headers)
-            self.assertEqual(response.status_code, 200) # Lock acquired
-
-            lock = PostLock.query.filter_by(post_id=self.test_post.id).first()
-            self.assertIsNotNone(lock)
-
-            # Check if api.socketio.emit was called with 'post_lock_acquired'
-            mock_api_socketio_emit.assert_called_with(
-                'post_lock_acquired',
-                {
-                    'post_id': self.test_post.id,
-                    'user_id': self.collaborator.id,
-                    'username': self.collaborator.username,
-                    'expires_at': lock.expires_at.isoformat()
-                },
-                room=f'post_{self.test_post.id}'
-            )
-
-    @patch('api.socketio.emit') # Mock socketio.emit in the api module
-    def test_socketio_lock_released_broadcast_from_api(self, mock_api_socketio_emit):
-        with app.app_context():
-            # Collaborator acquires lock first
-            self._create_db_lock(post_id=self.test_post.id, user_id=self.collaborator.id, minutes_offset=15)
-
-            token = self._get_jwt_token(self.collaborator.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.delete(f'/api/posts/{self.test_post.id}/lock', headers=headers)
-            self.assertEqual(response.status_code, 200) # Lock released
-
-            # Check if api.socketio.emit was called with 'post_lock_released'
-            mock_api_socketio_emit.assert_called_with(
-                'post_lock_released',
-                {
-                    'post_id': self.test_post.id,
-                    'released_by_user_id': self.collaborator.id,
-                    'username': self.collaborator.username
-                },
-                room=f'post_{self.test_post.id}'
-            )
-
-
-
-class TestRealtimePostNotifications(AppTestCase):
-
-    @patch('app.broadcast_new_post') # Patching the function in app.py
-    def test_create_post_api_triggers_broadcast(self, mock_broadcast_new_post_func):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password') # Now uses AppTestCase._get_jwt_token
-            headers = {
-                'Authorization': f'Bearer {token}',
-                'Content-Type': 'application/json'
-            }
-
-            post_payload = {
-                'title': 'Realtime Test Post API SSE',
-                'content': 'This API post should trigger a broadcast with a snippet for SSE.'
-            }
-
-            response = self.client.post('/api/posts', headers=headers, json=post_payload)
-
-            self.assertEqual(response.status_code, 201, f"API post creation failed: {response.data.decode()}")
-            response_json = response.get_json()
-            self.assertIn('post', response_json)
-            created_post_from_response = response_json['post']
-
-            mock_broadcast_new_post_func.assert_called_once()
-
-            args_call_list = mock_broadcast_new_post_func.call_args_list
-            self.assertEqual(len(args_call_list), 1)
-
-            called_with_args, called_with_kwargs = args_call_list[0]
-            self.assertEqual(len(called_with_args), 1)
-
-            broadcast_arg_dict = called_with_args[0]
-
-            self.assertEqual(broadcast_arg_dict['id'], created_post_from_response['id'])
-            self.assertEqual(broadcast_arg_dict['title'], post_payload['title'])
-
-            content_to_check = post_payload['content']
-            expected_snippet = (content_to_check[:100] + '...' if len(content_to_check) > 100 else content_to_check)
-            self.assertEqual(broadcast_arg_dict['content_snippet'], expected_snippet)
-
-            self.assertEqual(broadcast_arg_dict['author_username'], self.user1.username)
-            # Check that 'content' (full) and 'url' are NOT in the dict passed to broadcast_new_post
-            # because 'url' is added by broadcast_new_post itself.
-            self.assertNotIn('content', broadcast_arg_dict)
-            self.assertNotIn('url', broadcast_arg_dict)
-
-    def test_post_stream_endpoint_basic_connection(self):
-        with app.app_context():
-            response = self.client.get('/api/posts/stream')
-            self.assertEqual(response.status_code, 200)
-            self.assertEqual(response.mimetype, 'text/event-stream')
-            self.assertTrue(response.is_streamed, "Response from /api/posts/stream should be streamed.")
-            # Closing the response explicitly after checking it's streamed
-            response.close()
-
-
-class TestLikeNotifications(AppTestCase):
-
-    def setUp(self):
-        super().setUp()
-        # self.user1 (author), self.user2 (liker) are created by AppTestCase._setup_base_users()
-        self.author = self.user1
-        self.liker = self.user2
-
-    @patch('app.socketio.emit')
-    def test_like_post_sends_notification_and_emits_event(self, mock_socketio_emit):
-        with app.app_context():
-            # 1. Setup Post by author
-            post_by_author = self._create_db_post(user_id=self.author.id, title="Author's Likable Post")
-
-            # 2. Login as liker
-            self.login(self.liker.username, 'password')
-
-            # 3. Liker likes the post
-            response = self.client.post(f'/blog/post/{post_by_author.id}/like', follow_redirects=True)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(b"Post liked!", response.data)
-
-            # 4. Verify Database Notification for author
-            notification = Notification.query.filter_by(
-                user_id=self.author.id,
-                type="new_like",
-                related_id=post_by_author.id
-            ).first()
-            self.assertIsNotNone(notification, "Database notification was not created for the author.")
-            expected_message = f"{self.liker.username} liked your post: '{post_by_author.title}'"
-            self.assertEqual(notification.message, expected_message)
-
-            # 5. Verify SocketIO Emission to author's room
-            expected_payload = {
-                'liker_username': self.liker.username,
-                'post_id': post_by_author.id,
-                'post_title': post_by_author.title,
-                'message': expected_message,
-                'notification_id': notification.id
-            }
-            # Check if the specific call was made.
-            # Iterating through call_args_list is more robust if other emits might happen.
-            found_call = False
-            for call_args_tuple in mock_socketio_emit.call_args_list:
-                args, kwargs = call_args_tuple
-                if args[0] == 'new_like_notification' and kwargs.get('room') == f'user_{self.author.id}':
-                    self.assertEqual(args[1], expected_payload)
-                    found_call = True
-                    break
-            self.assertTrue(found_call, "Expected 'new_like_notification' event was not emitted to the author's room with correct payload.")
-
-            self.logout()
-
-    @patch('app.socketio.emit')
-    def test_like_own_post_does_not_send_notification_or_emit_event(self, mock_socketio_emit):
-        with app.app_context():
-            # 1. Setup Post by author
-            post_by_author = self._create_db_post(user_id=self.author.id, title="Author's Own Post to Like")
-
-            # 2. Login as author
-            self.login(self.author.username, 'password')
-
-            # 3. Author likes their own post
-            response = self.client.post(f'/blog/post/{post_by_author.id}/like', follow_redirects=True)
-            self.assertEqual(response.status_code, 200)
-            self.assertIn(b"Post liked!", response.data) # The like itself should still work
-
-            # 4. Verify NO Database Notification for author
-            notification = Notification.query.filter_by(
-                user_id=self.author.id,
-                type="new_like",
-                related_id=post_by_author.id
-            ).first()
-            self.assertIsNone(notification, "Notification should NOT be created when a user likes their own post.")
-
-            # 5. Verify NO SocketIO Emission of 'new_like_notification'
-            like_notification_event_found = False
-            for call_args_tuple in mock_socketio_emit.call_args_list:
-                args, kwargs = call_args_tuple
-                if args[0] == 'new_like_notification' and kwargs.get('room') == f'user_{self.author.id}':
-                    like_notification_event_found = True
-                    break
-            self.assertFalse(like_notification_event_found, "new_like_notification event should NOT be emitted when a user likes their own post.")
-
-            self.logout()
-
-
-def seed_test_achievements():
-    achievements_data = [
-        {"name": "Test First Post", "description": "Desc1", "icon_url": "icon1", "criteria_type": "num_posts", "criteria_value": 1},
-        {"name": "Test 5 Posts", "description": "Desc2", "icon_url": "icon2", "criteria_type": "num_posts", "criteria_value": 5},
-        {"name": "Test First Comment", "description": "Desc3", "icon_url": "icon3", "criteria_type": "num_comments_given", "criteria_value": 1},
-    ]
-    ach_ids = {}
-    for ach_data in achievements_data:
-        existing_achievement = Achievement.query.filter_by(name=ach_data["name"]).first()
-        if not existing_achievement:
-            ach = Achievement(**ach_data)
-            db.session.add(ach)
-            db.session.commit() # Commit after each add to get ID
-            ach_ids[ach_data['name']] = ach.id
-        else:
-            ach_ids[ach_data['name']] = existing_achievement.id
-
-    # Ensure all are committed before returning IDs based on fresh query
-    # This might be redundant if committing after each add, but safe.
-    db.session.commit()
-    # Re-fetch IDs to be certain, especially if some existed and some were added.
-    final_ach_ids = {ach_data['name']: Achievement.query.filter_by(name=ach_data['name']).first().id for ach_data in achievements_data}
-    return final_ach_ids
-
-class AchievementLogicTests(AppTestCase):
-
-    def test_get_user_stat_num_posts(self):
-        with app.app_context():
-            # user1 is created in AppTestCase's setUp
-            user = self.user1
-
-            # Add posts for the user
-            post1 = Post(title='P1', content='C1', user_id=user.id)
-            post2 = Post(title='P2', content='C2', user_id=user.id)
-            db.session.add_all([post1, post2])
-            db.session.commit()
-
-            self.assertEqual(get_user_stat(user, 'num_posts'), 2)
-            self.assertEqual(get_user_stat(user, 'num_comments_given'), 0)
-
-    def test_award_first_post_achievement(self):
-        with app.app_context():
-            seed_test_achievements()
-            user = self.user2 # Using a different user from AppTestCase
-
-            # Initially no achievement
-            self.assertEqual(UserAchievement.query.filter_by(user_id=user.id).count(), 0)
-
-            # User creates a post
-            post = Post(title='My First Test Post', content='Content', user_id=user.id)
-            db.session.add(post)
-            db.session.commit()
-
-            check_and_award_achievements(user.id)
-
-            first_post_ach = Achievement.query.filter_by(name="Test First Post").first()
-            self.assertIsNotNone(first_post_ach)
-            user_ach = UserAchievement.query.filter_by(user_id=user.id, achievement_id=first_post_ach.id).first()
-            self.assertIsNotNone(user_ach)
-
-            five_posts_ach = Achievement.query.filter_by(name="Test 5 Posts").first()
-            if five_posts_ach:
-                user_ach_5_posts = UserAchievement.query.filter_by(user_id=user.id, achievement_id=five_posts_ach.id).first()
-                self.assertIsNone(user_ach_5_posts)
-
-    def test_award_multiple_achievements_incrementally(self):
-        with app.app_context():
-            ach_ids = seed_test_achievements()
-            user = self.user3 # Using another user
-
-            # Create 1st post
-            db.session.add(Post(title='T1', content='C', user_id=user.id))
-            db.session.commit()
-            check_and_award_achievements(user.id)
-            self.assertIsNotNone(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test First Post"]).first())
-            if "Test 5 Posts" in ach_ids:
-                self.assertIsNone(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test 5 Posts"]).first())
-
-            for i in range(2, 5): # Posts 2, 3, 4
-                db.session.add(Post(title=f'T{i}', content='C', user_id=user.id))
-            db.session.commit()
-            check_and_award_achievements(user.id)
-            if "Test 5 Posts" in ach_ids:
-                self.assertIsNone(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test 5 Posts"]).first())
-
-            db.session.add(Post(title='T5', content='C', user_id=user.id)) # 5th post
-            db.session.commit()
-            check_and_award_achievements(user.id)
-            if "Test 5 Posts" in ach_ids:
-                self.assertIsNotNone(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test 5 Posts"]).first())
-
-    def test_no_duplicate_achievements_awarded(self):
-        with app.app_context():
-            ach_ids = seed_test_achievements()
-            user = self.user1
-
-            db.session.add(Post(title='P1 Duplicate Test', content='C', user_id=user.id))
-            db.session.commit()
-
-            check_and_award_achievements(user.id)
-            self.assertEqual(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test First Post"]).count(), 1)
-
-            check_and_award_achievements(user.id) # Call again
-            self.assertEqual(UserAchievement.query.filter_by(user_id=user.id, achievement_id=ach_ids["Test First Post"]).count(), 1)
-
-    def test_display_achievements_on_user_profile(self):
-        with app.app_context():
-            ach_ids = seed_test_achievements()
-            user = User(username='profile_ach_user', email='pau@example.com', password_hash=generate_password_hash('password123'))
-            db.session.add(user)
-            db.session.commit()
-
-            first_post_ach_id = ach_ids.get("Test First Post")
-            self.assertIsNotNone(first_post_ach_id, "Test First Post achievement was not seeded correctly.")
-            user_ach = UserAchievement(user_id=user.id, achievement_id=first_post_ach_id)
-            db.session.add(user_ach)
-            db.session.commit()
-
-            response = self.client.get(f'/user/{user.username}')
-            self.assertEqual(response.status_code, 200)
-
-            response_data = response.get_data(as_text=True)
-            # Check for the achievements section title
-            self.assertIn("<h3>Achievements</h3>", response_data)
-            # Check if earned achievement name is present
-            self.assertIn("Test First Post", response_data)
-            self.assertIn("[ICON]", response_data)
-
-            # Check that an unearned achievement (e.g. Test 5 Posts) is NOT directly listed in the earned section
-            # This is a bit tricky as "Test 5 Posts" might appear in a link to "All Achievements"
-            # We need to be more specific if the test fails due to that.
-            # For now, assume it's not part of a general link text on the profile page itself.
-            if "Test 5 Posts" in ach_ids:
-                # A more robust check would parse HTML, but this is a basic check.
-                # We are checking if "Test 5 Posts" appears specifically within the rendered achievement items.
-                # The current structure in user.html for achievements is:
-                # <div class="achievement-item" title="..."> <span class="achievement-icon">...</span> <p class="achievement-name">ACH_NAME</p> </div>
-                # So, if "Test 5 Posts" is NOT earned, it should not appear as a <p class="achievement-name">Test 5 Posts</p>
-                self.assertNotIn('<p class="achievement-name">Test 5 Posts</p>', response_data)
-
-    def test_no_achievements_message_on_profile(self):
-        with app.app_context():
-            user = User(username='no_ach_user_profile', email='naup@example.com', password_hash=generate_password_hash('password123'))
-            db.session.add(user)
-            db.session.commit()
-
-            response = self.client.get(f'/user/{user.username}')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-            self.assertIn("hasn&#39;t earned any achievements yet.</p>", response_data) # Check for HTML escaped apostrophe
-            self.assertNotIn('<div class="achievements-grid">', response_data) # Grid should not be there if no achievements
-
-    def test_view_user_achievements_page_earned_and_all(self):
-        with app.app_context():
-            ach_ids = seed_test_achievements()
-            user = User(username='all_ach_user_page', email='aaup@example.com', password_hash=generate_password_hash('password123'))
-            db.session.add(user)
-            db.session.commit()
-
-            first_comment_ach_id = ach_ids.get("Test First Comment")
-            self.assertIsNotNone(first_comment_ach_id, "Test First Comment achievement not seeded.")
-            user_ach = UserAchievement(user_id=user.id, achievement_id=first_comment_ach_id)
-            db.session.add(user_ach)
-            db.session.commit()
-
-            self.login(user.username, 'password123')
-
-            response = self.client.get(f'/user/{user.username}/achievements')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            self.assertIn(f"{user.username}'s Achievements", response_data)
-
-            self.assertIn("<h3>Earned Achievements", response_data)
-            self.assertIn("Test First Comment", response_data)
-            self.assertIn("Awarded:", response_data)
-
-            self.assertIn("<h3>All Available Achievements</h3>", response_data)
-            self.assertIn("Test First Post", response_data)
-            self.assertIn("Test 5 Posts", response_data)
-
-            # Using more specific checks for badges based on provided HTML structure
-            self.assertIn('Test First Comment</h5>\n                    <p class="mb-1">Desc3</p>\n                    <small>Awarded: ', response_data) # Check it's in the earned list detail
-            self.assertIn('Test First Comment\n                       </h5>\n                       <p class="mb-1">Desc3</p>\n                       <small class="text-muted">Criteria: num_comments_given >= 1</small>', response_data) # Check this part to ensure it's in the "All" list
-            # The above is too fragile. Let's check for the badge directly.
-            self.assertInHTML('<span class="badge bg-success float-end">Earned</span>', response_data, achievement_name="Test First Comment")
-            self.assertInHTML('<span class="badge bg-secondary float-end">Not Earned</span>', response_data, achievement_name="Test First Post")
-            self.logout()
-
-    def test_view_user_achievements_page_no_earned(self):
-        with app.app_context():
-            seed_test_achievements()
-            user = User(username='no_earned_ach_page', email='neaup@example.com', password_hash=generate_password_hash('password123'))
-            db.session.add(user)
-            db.session.commit()
-
-            self.login(user.username, 'password123')
-
-            response = self.client.get(f'/user/{user.username}/achievements')
-            self.assertEqual(response.status_code, 200)
-            response_data = response.get_data(as_text=True)
-
-            self.assertIn(f"{user.username}'s Achievements", response_data)
-            self.assertIn(f"{user.username} has not earned any achievements yet.</p>", response_data)
-
-            self.assertIn("<h3>All Available Achievements</h3>", response_data)
-            self.assertIn("Test First Post", response_data)
-            self.assertIn("badge bg-secondary float-end", response_data)
-            self.assertNotIn("badge bg-success float-end", response_data)
-            self.logout()
-
-    # Helper for more robust HTML checking, if needed, can be added to AppTestCase
-    def assertInHTML(self, needle, haystack, achievement_name):
-        # This is a simplified check. For real HTML parsing, use BeautifulSoup or lxml.
-        # It tries to find the achievement name and then the needle (badge) in its vicinity.
-        try:
-            ach_name_idx = haystack.find(achievement_name)
-            self.assertTrue(ach_name_idx != -1, f"Achievement name '{achievement_name}' not found in HTML.")
-            # Search for the needle (e.g., badge HTML) within a reasonable range after the achievement name.
-            # This range might need adjustment based on actual HTML structure.
-            search_range = haystack[ach_name_idx : ach_name_idx + 300]
-            self.assertIn(needle, search_range, f"'{needle}' not found near '{achievement_name}' in HTML: {search_range[:200]}...")
-        except AttributeError: # if haystack is not a string
-             self.fail("Haystack for assertInHTML was not a string.")
-
-    def test_delete_shared_file_sender(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password') # Sender
-            dummy_file_data = self.create_dummy_file(filename="to_delete_sender.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="to_delete_sender.txt").first()
-            self.assertIsNotNone(shared_file)
-            saved_filename = shared_file.saved_filename
-            file_id = shared_file.id
-            file_path = os.path.join(app.config['SHARED_FILES_UPLOAD_FOLDER'], saved_filename)
-            self.assertTrue(os.path.exists(file_path))
-            # Do not logout, sender is already logged in
-
-            response = self.client.post(f'/files/delete/{file_id}', follow_redirects=True)
-            self.assertEqual(response.status_code, 200) # After redirect to inbox
-            self.assertIn("File successfully deleted.", response.get_data(as_text=True))
-
-            self.assertIsNone(SharedFile.query.get(file_id))
-            self.assertFalse(os.path.exists(file_path))
-            self.logout()
-
-    def test_delete_shared_file_unauthorized(self):
-        with app.app_context():
-            self.login(self.user1.username, 'password')
-            dummy_file_data = self.create_dummy_file(filename="unauth_delete.txt")
-            self.client.post(f'/files/share/{self.user2.username}', data={'file': dummy_file_data}, content_type='multipart/form-data')
-            shared_file = SharedFile.query.filter_by(original_filename="unauth_delete.txt").first()
-            self.assertIsNotNone(shared_file)
-            file_id = shared_file.id
-            file_path = os.path.join(app.config['SHARED_FILES_UPLOAD_FOLDER'], shared_file.saved_filename)
-            self.logout()
-
-            self.login(self.user3.username, 'password') # Unauthorized user
-            response = self.client.post(f'/files/delete/{file_id}', follow_redirects=True)
-            self.assertEqual(response.status_code, 200) # After redirect
-            self.assertIn("You are not authorized to delete this file.", response.get_data(as_text=True))
-
-            self.assertIsNotNone(SharedFile.query.get(file_id)) # Still exists
-            self.assertFalse(os.path.exists(file_path)) # File should be gone
-            self.logout()
-
-
-
-
-class TestSeriesFeature(AppTestCase):
-    # Model Tests
-    def test_series_model_creation(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="My First Series", description="Test description.")
-            self.assertIsNotNone(series.id)
-            self.assertEqual(series.title, "My First Series")
-            self.assertEqual(series.author, self.user1)
-            self.assertIn(series, self.user1.series_created)
-
-    def test_series_post_association_and_order(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            post1 = self._create_db_post(user_id=self.user1_id, title="Post 1 for Series")
-            post2 = self._create_db_post(user_id=self.user1_id, title="Post 2 for Series")
-
-            sp1 = SeriesPost(series_id=series.id, post_id=post1.id, order=1)
-            sp2 = SeriesPost(series_id=series.id, post_id=post2.id, order=2)
-            db.session.add_all([sp1, sp2])
-            db.session.commit()
-
-            # Refresh series to ensure 'posts' relationship is loaded with new data
-            db.session.refresh(series)
-
-            self.assertEqual(len(series.posts), 2)
-            self.assertEqual(series.posts[0].id, post1.id)
-            self.assertEqual(series.posts[1].id, post2.id)
-
-            assoc1 = SeriesPost.query.filter_by(series_id=series.id, post_id=post1.id).first()
-            self.assertIsNotNone(assoc1)
-            self.assertEqual(assoc1.order, 1)
-
-    def test_cascade_delete_user_to_series(self):
-        with app.app_context():
-            user_to_delete = User(username="deleteme_series", email="del@series.com", password_hash=generate_password_hash("pw"))
-            db.session.add(user_to_delete)
-            db.session.commit()
-            # Re-fetch to ensure it's in the session before creating series
-            user_to_delete_from_session = User.query.filter_by(username="deleteme_series").first()
-
-
-            series = self._create_series(user_id=user_to_delete_from_session.id, title="Series to be deleted")
-            series_id = series.id
-
-            db.session.delete(user_to_delete_from_session)
-            db.session.commit()
-
-            self.assertIsNone(Series.query.get(series_id))
-
-    def test_cascade_delete_series_to_series_post(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            post = self._create_db_post(user_id=self.user1_id)
-            sp = SeriesPost(series_id=series.id, post_id=post.id, order=1)
-            db.session.add(sp)
-            db.session.commit()
-            sp_id_tuple = (series.id, post.id)
-
-            self.assertIsNotNone(SeriesPost.query.get(sp_id_tuple))
-
-            db.session.delete(series)
-            db.session.commit()
-            self.assertIsNone(SeriesPost.query.get(sp_id_tuple))
-
-    def test_cascade_delete_post_to_series_post(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            post_to_delete = self._create_db_post(user_id=self.user1_id)
-            post_id = post_to_delete.id
-            sp = SeriesPost(series_id=series.id, post_id=post_id, order=1)
-            db.session.add(sp)
-            db.session.commit()
-            sp_id_tuple = (series.id, post_id)
-
-            self.assertIsNotNone(SeriesPost.query.get(sp_id_tuple))
-
-            db.session.delete(post_to_delete)
-            db.session.commit()
-            self.assertIsNone(SeriesPost.query.get(sp_id_tuple))
-
-    # --- Route Tests ---
-    def test_create_series_page_load(self):
-        self.login(self.user1.username, 'password')
-        response = self.client.get('/series/create')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Create New Series', response.data)
-        self.logout()
-
-    def test_create_series_unauthenticated(self):
-        response = self.client.get('/series/create', follow_redirects=False)
-        self.assertEqual(response.status_code, 302)
-        self.assertIn('/login', response.location)
-
-        response_post = self.client.post('/series/create', data={'title': 'Fail Series'}, follow_redirects=False)
-        self.assertEqual(response_post.status_code, 302)
-        self.assertIn('/login', response_post.location)
-
-    def test_create_series_post_success(self):
-        self.login(self.user1.username, 'password')
-        response = self.client.post('/series/create', data={
-            'title': 'My Awesome Series',
-            'description': 'A description of awesomeness.'
-        }, follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Series created successfully!', response.data)
-        self.assertIn(b'My Awesome Series', response.data)
-
-        with app.app_context():
-            series = Series.query.filter_by(title='My Awesome Series').first()
-            self.assertIsNotNone(series)
-            self.assertEqual(series.user_id, self.user1_id)
-            self.assertEqual(series.description, 'A description of awesomeness.')
-        self.logout()
-
-    def test_create_series_post_no_title(self):
-        self.login(self.user1.username, 'password')
-        response = self.client.post('/series/create', data={
-            'title': '',
-            'description': 'This should fail.'
-        }, follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Series title cannot be empty.', response.data)
-        self.assertIn(b'Create New Series', response.data)
-
-        with app.app_context():
-            self.assertIsNone(Series.query.filter_by(description='This should fail.').first())
-        self.logout()
-
-    def test_view_series_page(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Viewable Series", description="Desc")
-            post1 = self._create_db_post(user_id=self.user1_id, title="Post In Series")
-            sp1 = SeriesPost(series_id=series.id, post_id=post1.id, order=1)
-            db.session.add(sp1)
-            db.session.commit()
-
-        response = self.client.get(f'/series/{series.id}')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Viewable Series', response.data)
-        self.assertIn(b'Desc', response.data)
-        self.assertIn(b'Post In Series', response.data)
-        self.assertIn(b'#1', response.data)
-
-    def test_view_series_not_found(self):
-        response = self.client.get('/series/9999')
-        self.assertEqual(response.status_code, 404)
-
-    def test_edit_series_page_load_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Editable Series")
-
-        self.login(self.user1.username, 'password')
-        response = self.client.get(f'/series/{series.id}/edit')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Edit Series: Editable Series', response.data)
-        self.assertIn(b'value="Editable Series"', response.data)
-        self.logout()
-
-    def test_edit_series_page_load_non_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="NonEditable Series")
-
-        self.login(self.user2.username, 'password')
-        response = self.client.get(f'/series/{series.id}/edit', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'You are not authorized to edit this series.', response.data)
-        self.assertIn(b'NonEditable Series', response.data)
-        self.logout()
-
-    def test_edit_series_post_success_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Original Title")
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series.id}/edit', data={
-            'title': 'Updated Title',
-            'description': 'Updated Description.'
-        }, follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Series details updated successfully!', response.data)
-        self.assertIn(b'Edit Series: Updated Title', response.data)
-
-        with app.app_context():
-            updated_series = Series.query.get(series.id)
-            self.assertEqual(updated_series.title, 'Updated Title')
-            self.assertEqual(updated_series.description, 'Updated Description.')
-        self.logout()
-
-    def test_edit_series_post_non_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Cannot Edit")
-
-        self.login(self.user2.username, 'password')
-        response = self.client.post(f'/series/{series.id}/edit', data={
-            'title': 'Attempted Update'
-        }, follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'You are not authorized to edit this series.', response.data)
-
-        with app.app_context():
-            self.assertEqual(Series.query.get(series.id).title, 'Cannot Edit')
-        self.logout()
-
-    def test_delete_series_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="To Be Deleted")
-            post = self._create_db_post(user_id=self.user1_id)
-            sp = SeriesPost(series_id=series.id, post_id=post.id, order=1)
-            db.session.add(sp)
-            db.session.commit()
-            series_id = series.id
-            sp_id_tuple = (series.id, post.id)
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series_id}/delete', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Series deleted successfully.', response.data)
-        self.assertIn(f"{self.user1.username}'s Profile".encode(), response.data)
-
-        with app.app_context():
-            self.assertIsNone(Series.query.get(series_id))
-            self.assertIsNone(SeriesPost.query.get(sp_id_tuple))
-        self.logout()
-
-    def test_delete_series_non_author(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Cannot Delete")
-            series_id = series.id
-
-        self.login(self.user2.username, 'password')
-        response = self.client.post(f'/series/{series_id}/delete', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'You are not authorized to delete this series.', response.data)
-
-        with app.app_context():
-            self.assertIsNotNone(Series.query.get(series_id))
-        self.logout()
-
-    # --- Manage Posts in Series Route Tests ---
-    def test_add_post_to_series_success(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            post_to_add = self._create_db_post(user_id=self.user1_id, title="Post To Add")
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series.id}/add_post/{post_to_add.id}', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(f"Post '{post_to_add.title}' added to series '{series.title}'.".encode(), response.data)
-
-        with app.app_context():
-            sp_entry = SeriesPost.query.filter_by(series_id=series.id, post_id=post_to_add.id).first()
-            self.assertIsNotNone(sp_entry)
-            self.assertEqual(sp_entry.order, 1)
-        self.logout()
-
-    def test_add_post_to_series_already_exists(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            post = self._create_db_post(user_id=self.user1_id)
-            sp = SeriesPost(series_id=series.id, post_id=post.id, order=1)
-            db.session.add(sp)
-            db.session.commit()
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series.id}/add_post/{post.id}', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'This post is already in the series.', response.data)
-        self.logout()
-
-    def test_add_post_to_series_non_author_of_series(self):
-        with app.app_context():
-            series_by_user1 = self._create_series(user_id=self.user1_id)
-            post_by_user2 = self._create_db_post(user_id=self.user2_id, title="User2 Post")
-
-        self.login(self.user2.username, 'password')
-        response = self.client.post(f'/series/{series_by_user1.id}/add_post/{post_by_user2.id}', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'You are not authorized to modify this series.', response.data)
-
-        with app.app_context():
-            self.assertIsNone(SeriesPost.query.filter_by(series_id=series_by_user1.id, post_id=post_by_user2.id).first())
-        self.logout()
-
-    def test_add_post_to_series_post_not_owned_by_series_author(self):
-        with app.app_context():
-            series_by_user1 = self._create_series(user_id=self.user1_id)
-            post_by_user2 = self._create_db_post(user_id=self.user2_id, title="User2 Post")
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series_by_user1.id}/add_post/{post_by_user2.id}', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'You can only add your own posts to your series.', response.data)
-
-        with app.app_context():
-            self.assertIsNone(SeriesPost.query.filter_by(series_id=series_by_user1.id, post_id=post_by_user2.id).first())
-        self.logout()
-
-    def test_remove_post_from_series_and_reorder(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id)
-            p1 = self._create_db_post(user_id=self.user1_id, title="P1")
-            p2 = self._create_db_post(user_id=self.user1_id, title="P2")
-            p3 = self._create_db_post(user_id=self.user1_id, title="P3")
-            db.session.add_all([
-                SeriesPost(series_id=series.id, post_id=p1.id, order=1),
-                SeriesPost(series_id=series.id, post_id=p2.id, order=2),
-                SeriesPost(series_id=series.id, post_id=p3.id, order=3)
-            ])
-            db.session.commit()
-
-        self.login(self.user1.username, 'password')
-        response = self.client.post(f'/series/{series.id}/remove_post/{p2.id}', follow_redirects=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(f"Post '{p2.title}' removed from series '{series.title}'.".encode(), response.data)
-
-        with app.app_context():
-            self.assertIsNone(SeriesPost.query.filter_by(series_id=series.id, post_id=p2.id).first())
-            sp1 = SeriesPost.query.filter_by(series_id=series.id, post_id=p1.id).first()
-            sp3 = SeriesPost.query.filter_by(series_id=series.id, post_id=p3.id).first()
-            self.assertIsNotNone(sp1)
-            self.assertIsNotNone(sp3)
-            self.assertEqual(sp1.order, 1)
-            self.assertEqual(sp3.order, 2)
-        self.logout()
-
-    # --- UI/Content Tests (Simplified) ---
-    def test_user_profile_lists_series(self):
-        with app.app_context():
-            series1 = self._create_series(user_id=self.user1_id, title="User1 Series One")
-            series2 = self._create_series(user_id=self.user1_id, title="User1 Series Two")
-
-        response = self.client.get(f'/user/{self.user1.username}')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'User1 Series One', response.data)
-        self.assertIn(b'User1 Series Two', response.data)
-        self.assertIn(f'href="/series/{series1.id}"'.encode(), response.data)
-
-    def test_view_post_lists_series_and_navigation(self):
-        with app.app_context():
-            series = self._create_series(user_id=self.user1_id, title="Nav Series")
-            p1 = self._create_db_post(user_id=self.user1_id, title="Nav Post 1")
-            p2 = self._create_db_post(user_id=self.user1_id, title="Nav Post 2")
-            p3 = self._create_db_post(user_id=self.user1_id, title="Nav Post 3")
-            db.session.add_all([
-                SeriesPost(series_id=series.id, post_id=p1.id, order=1),
-                SeriesPost(series_id=series.id, post_id=p2.id, order=2),
-                SeriesPost(series_id=series.id, post_id=p3.id, order=3)
-            ])
-            db.session.commit()
-
-        # Test middle post (p2)
-        response = self.client.get(f'/blog/post/{p2.id}?series_id={series.id}')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Part of Series:', response.data)
-        self.assertIn(f'href="/series/{series.id}"'.encode(), response.data)
-        self.assertIn(b'Nav Series', response.data)
-        self.assertIn(b'Previous in series: Nav Post 1', response.data)
-        self.assertIn(f'href="/blog/post/{p1.id}?series_id={series.id}"'.encode(), response.data)
-        self.assertIn(b'Next in series: Nav Post 3', response.data)
-        self.assertIn(f'href="/blog/post/{p3.id}?series_id={series.id}"'.encode(), response.data)
-
-        # Test first post (p1)
-        response = self.client.get(f'/blog/post/{p1.id}?series_id={series.id}')
-        self.assertEqual(response.status_code, 200)
-        self.assertNotIn(b'Previous in series', response.data)
-        self.assertIn(b'Next in series: Nav Post 2', response.data)
-
-        # Test last post (p3)
-        response = self.client.get(f'/blog/post/{p3.id}?series_id={series.id}')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Previous in series: Nav Post 2', response.data)
-        self.assertNotIn(b'Next in series', response.data)
-
-        # Test without series_id in query (no prev/next links)
-        response = self.client.get(f'/blog/post/{p2.id}')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Part of Series:', response.data)
-        self.assertNotIn(b'Previous in series', response.data)
-        self.assertNotIn(b'Next in series', response.data)
-
-
-class TestCommentAPI(AppTestCase):
-
-    def test_create_comment_success(self):
-        with app.app_context():
-            # 1. Create a test user and a test post.
-            # user1 is created in AppTestCase's setUp
-            test_post = self._create_db_post(user_id=self.user1_id, title="Post for Commenting")
-            post_id = test_post.id
-
-            # 2. Log in as the test user to get a JWT token.
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {
-                'Authorization': f'Bearer {token}',
-                'Content-Type': 'application/json'
-            }
-
-            # 3. Make a POST request to /api/posts/{post_id}/comments
-            comment_content = "This is a test comment."
-            response = self.client.post(f'/api/posts/{post_id}/comments', headers=headers, json={'content': comment_content})
-
-            # 4. Assert that the response status code is 201.
-            self.assertEqual(response.status_code, 201, f"Response data: {response.data.decode()}")
-
-            # 5. Assert that the response JSON contains the correct message and comment details.
-            data = json.loads(response.data)
-            self.assertEqual(data['message'], 'Comment created successfully')
-            self.assertIn('comment', data)
-            comment_data = data['comment']
-            self.assertEqual(comment_data['content'], comment_content)
-            self.assertEqual(comment_data['user_id'], self.user1_id)
-            self.assertEqual(comment_data['author_username'], self.user1.username)
-            self.assertEqual(comment_data['post_id'], post_id)
-            self.assertIsNotNone(comment_data['id'])
-            self.assertIsNotNone(comment_data['timestamp'])
-
-            # 6. Verify that the comment is actually created in the database.
-            comment_in_db = Comment.query.get(comment_data['id'])
-            self.assertIsNotNone(comment_in_db)
-            self.assertEqual(comment_in_db.content, comment_content)
-            self.assertEqual(comment_in_db.user_id, self.user1_id)
-            self.assertEqual(comment_in_db.post_id, post_id)
-
-    def test_create_comment_unauthenticated(self):
-        with app.app_context():
-            test_post = self._create_db_post(user_id=self.user1_id, title="Post for Unauth Comment")
-            post_id = test_post.id
-
-            # Make request WITHOUT token
-            headers = {'Content-Type': 'application/json'}
-            response = self.client.post(f'/api/posts/{post_id}/comments', headers=headers, json={'content': "A comment attempt"})
-
-            self.assertEqual(response.status_code, 401) # Expecting 401 for missing JWT
-
-    def test_create_comment_post_not_found(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {
-                'Authorization': f'Bearer {token}',
-                'Content-Type': 'application/json'
-            }
-            non_existent_post_id = 99999
-
-            response = self.client.post(f'/api/posts/{non_existent_post_id}/comments', headers=headers, json={'content': "Commenting on nothing"})
-
-            self.assertEqual(response.status_code, 404)
-            data = json.loads(response.data)
-            self.assertEqual(data['message'], 'Post not found')
-
-    def test_create_comment_missing_content(self):
-        with app.app_context():
-            test_post = self._create_db_post(user_id=self.user1_id, title="Post for Invalid Comment")
-            post_id = test_post.id
-
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {
-                'Authorization': f'Bearer {token}',
-                'Content-Type': 'application/json'
-            }
-
-            response = self.client.post(f'/api/posts/{post_id}/comments', headers=headers, json={}) # Empty JSON body
-
-            self.assertEqual(response.status_code, 400)
-            data = json.loads(response.data)
-            # The exact message depends on reqparse error formatting.
-            # It usually returns a dict where keys are problematic arguments.
-            self.assertIn('content', data['message'])
-            self.assertIn('cannot be blank', data['message']['content'].lower())
-
-    # Helper for creating polls in DB for test setup
-    def _create_db_poll(self, user_id, question="Test Poll?", options_texts=None, created_at=None):
-        if options_texts is None:
-            options_texts = ["Option 1", "Option 2"]
-        poll = Poll(question=question, user_id=user_id, created_at=created_at or datetime.utcnow())
-        db.session.add(poll)
-        db.session.flush()  # So poll gets an ID before adding options
-        for text in options_texts:
-            option = PollOption(text=text, poll_id=poll.id)
-            db.session.add(option)
-        db.session.commit()
-        return poll
-
-    # Helper for creating poll votes in DB
-    def _create_db_poll_vote(self, user_id, poll_id, poll_option_id, created_at=None):
-        # Note: PollVote model in the provided models.py doesn't have a timestamp/created_at.
-        # If it were added, it would be: timestamp=created_at or datetime.utcnow()
-        vote = PollVote(user_id=user_id, poll_id=poll_id, poll_option_id=poll_option_id)
-        db.session.add(vote)
-        db.session.commit()
-        return vote
-
-
-class TestPollAPI(AppTestCase):
-
-    def test_create_poll_success(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-            poll_data = {'question': 'What is your favorite color?', 'options': ['Red', 'Green', 'Blue']}
-
-            response = self.client.post('/api/polls', headers=headers, json=poll_data)
-            self.assertEqual(response.status_code, 201)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Poll created successfully')
-            self.assertIn('poll', data)
-            self.assertEqual(data['poll']['question'], poll_data['question'])
-            self.assertEqual(len(data['poll']['options']), 3)
-            self.assertEqual(data['poll']['author_username'], self.user1.username)
-
-            # Verify in DB
-            poll_in_db = Poll.query.get(data['poll']['id'])
-            self.assertIsNotNone(poll_in_db)
-            self.assertEqual(poll_in_db.question, poll_data['question'])
-            self.assertEqual(len(poll_in_db.options), 3)
-            self.assertEqual(poll_in_db.options[0].text, 'Red')
-
-    def test_create_poll_missing_data(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-
-            # Missing question
-            response = self.client.post('/api/polls', headers=headers, json={'options': ['Yes', 'No']})
-            self.assertEqual(response.status_code, 400)
-            data = response.get_json()
-            self.assertIn('question', data['message']) # reqparse error format
-
-            # Missing options
-            response = self.client.post('/api/polls', headers=headers, json={'question': 'Is this a test?'})
-            self.assertEqual(response.status_code, 400)
-            data = response.get_json()
-            self.assertIn('options', data['message'])
-
-    def test_create_poll_too_few_options(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-            poll_data = {'question': 'Need more options?', 'options': ['Just one']}
-
-            response = self.client.post('/api/polls', headers=headers, json=poll_data)
-            self.assertEqual(response.status_code, 400)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'A poll must have at least two options')
-
-    def test_create_poll_unauthenticated(self):
-        with app.app_context():
-            headers = {'Content-Type': 'application/json'}
-            poll_data = {'question': 'Who can post this?', 'options': ['Me', 'You']}
-            response = self.client.post('/api/polls', headers=headers, json=poll_data)
-            self.assertEqual(response.status_code, 401) # JWT Missing
-
-    def test_list_polls_success(self):
-        with app.app_context():
-            self._create_db_poll(user_id=self.user1.id, question="Poll 1")
-            self._create_db_poll(user_id=self.user2.id, question="Poll 2")
-
-            token = self._get_jwt_token(self.user1.username, 'password') # Listing requires auth
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.get('/api/polls', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertIn('polls', data)
-            self.assertEqual(len(data['polls']), 2)
-            self.assertEqual(data['polls'][0]['question'], "Poll 1") # Assuming default order by ID or creation time (Poll 1 older)
-
-    def test_list_polls_empty(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.get('/api/polls', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertIn('polls', data)
-            self.assertEqual(len(data['polls']), 0)
-
-    def test_get_poll_success(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id, question="Specific Poll")
-            token = self._get_jwt_token(self.user1.username, 'password') # Get requires auth
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.get(f'/api/polls/{poll.id}', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertIn('poll', data)
-            self.assertEqual(data['poll']['question'], "Specific Poll")
-            self.assertEqual(data['poll']['id'], poll.id)
-
-    def test_get_poll_not_found(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.get('/api/polls/99999', headers=headers) # Non-existent ID
-            self.assertEqual(response.status_code, 404)
-
-    def test_delete_poll_success(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id, question="To Be Deleted")
-            poll_id = poll.id
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-
-            response = self.client.delete(f'/api/polls/{poll_id}', headers=headers)
-            self.assertEqual(response.status_code, 200)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Poll deleted')
-            self.assertIsNone(Poll.query.get(poll_id))
-
-    def test_delete_poll_unauthorized(self):
-        with app.app_context():
-            poll_by_user1 = self._create_db_poll(user_id=self.user1.id, question="User1's Poll")
-            poll_id = poll_by_user1.id
-
-            # Log in as user2
-            token_user2 = self._get_jwt_token(self.user2.username, 'password')
-            headers = {'Authorization': f'Bearer {token_user2}'}
-
-            response = self.client.delete(f'/api/polls/{poll_id}', headers=headers)
-            self.assertEqual(response.status_code, 403) # Forbidden
-            data = response.get_json()
-            self.assertEqual(data['message'], 'You are not authorized to delete this poll')
-            self.assertIsNotNone(Poll.query.get(poll_id)) # Still in DB
-
-    def test_delete_poll_unauthenticated(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id, question="Unauth Delete Test")
-            poll_id = poll.id
-            response = self.client.delete(f'/api/polls/{poll_id}') # No token
-            self.assertEqual(response.status_code, 401)
-            self.assertIsNotNone(Poll.query.get(poll_id))
-
-    def test_delete_poll_not_found(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}'}
-            response = self.client.delete('/api/polls/99999', headers=headers) # Non-existent ID
-            self.assertEqual(response.status_code, 404)
-
-    def test_vote_on_poll_success(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id, question="Vote Test Poll", options_texts=['OptA', 'OptB'])
-            option_to_vote_on = poll.options[0]
-            token = self._get_jwt_token(self.user2.username, 'password') # user2 votes
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-
-            response = self.client.post(f'/api/polls/{poll.id}/vote', headers=headers, json={'option_id': option_to_vote_on.id})
-            self.assertEqual(response.status_code, 201)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Vote cast successfully')
-
-            vote_in_db = PollVote.query.filter_by(user_id=self.user2.id, poll_id=poll.id, poll_option_id=option_to_vote_on.id).first()
-            self.assertIsNotNone(vote_in_db)
-
-    def test_vote_on_poll_already_voted(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id)
-            option1 = poll.options[0]
-            option2 = poll.options[1]
-            # user2 already voted for option1
-            self._create_db_poll_vote(user_id=self.user2.id, poll_id=poll.id, poll_option_id=option1.id)
-
-            token = self._get_jwt_token(self.user2.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-            # Try to vote again (for option2 or option1)
-            response = self.client.post(f'/api/polls/{poll.id}/vote', headers=headers, json={'option_id': option2.id})
-            self.assertEqual(response.status_code, 400)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'You have already voted on this poll')
-
-    def test_vote_on_poll_invalid_option_id_for_poll(self):
-        with app.app_context():
-            poll1 = self._create_db_poll(user_id=self.user1.id, question="Poll 1")
-            # poll2 = self._create_db_poll(user_id=self.user1.id, question="Poll 2") # Not needed for this specific variant
-
-            token = self._get_jwt_token(self.user2.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-
-            non_existent_option_id = 9999
-            response = self.client.post(f'/api/polls/{poll1.id}/vote', headers=headers, json={'option_id': non_existent_option_id})
-            self.assertEqual(response.status_code, 404) # Option not found for this poll
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Poll option not found or does not belong to this poll')
-
-    def test_vote_on_poll_option_not_in_specific_poll(self):
-        with app.app_context():
-            poll1 = self._create_db_poll(user_id=self.user1.id, question="Poll One", options_texts=["P1Opt1"])
-            poll2 = self._create_db_poll(user_id=self.user1.id, question="Poll Two", options_texts=["P2Opt1"])
-            option_from_poll2 = poll2.options[0]
-
-            token = self._get_jwt_token(self.user2.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-
-            # Try to vote on poll1 using an option from poll2
-            response = self.client.post(f'/api/polls/{poll1.id}/vote', headers=headers, json={'option_id': option_from_poll2.id})
-            self.assertEqual(response.status_code, 404) # Option does not belong to poll1
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Poll option not found or does not belong to this poll')
-
-    def test_vote_on_poll_unauthenticated(self):
-        with app.app_context():
-            poll = self._create_db_poll(user_id=self.user1.id)
-            option_id = poll.options[0].id
-            headers = {'Content-Type': 'application/json'}
-            response = self.client.post(f'/api/polls/{poll.id}/vote', headers=headers, json={'option_id': option_id})
-            self.assertEqual(response.status_code, 401)
-
-    def test_vote_on_poll_non_existent_poll(self):
-        with app.app_context():
-            token = self._get_jwt_token(self.user1.username, 'password')
-            headers = {'Authorization': f'Bearer {token}', 'Content-Type': 'application/json'}
-            response = self.client.post('/api/polls/99999/vote', headers=headers, json={'option_id': 1}) # Non-existent poll
-            self.assertEqual(response.status_code, 404)
-            data = response.get_json()
-            self.assertEqual(data['message'], 'Poll not found')
-
-
-# Helper function to seed achievements for tests
+# All other test classes and helper functions have been moved.
+# If this file is run, only TestMinimalSanityCheck will execute from here.
+# Consider removing this file or keeping it as a placeholder if all tests are modularized.
